+++ .github/workflows/test-ubuntu.yml
@@ -23,5 +23,6 @@ jobs:
           sudo apt install -y xvfb x11-utils libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 xdotool
           sudo apt-get update
           sudo apt-get install ffmpeg libsm6 libxext6  -y
+          sudo apt install libegl1
       - name: Test with pytest
         run: ./tests/test.sh
