### 2023-09-11 26cc5be
add libegl1 to test dependencies
