### 2023-09-11 464c526
add dummy test; test fault handler in viz tests
