### 2023-09-11 f625322
add DOIs to references in paper.bib
