+++ paper/paper.bib
@@ -7,6 +7,7 @@
   pages={35--53},
   year={2010},
   publisher={MIT press}
+  doi={10.1162/pres.19.1.35}
 }
 
 @article{esch2018mne,
@@ -17,6 +18,7 @@
   pages={55--67},
   year={2018},
   publisher={Elsevier}
+  doi={10.1016/j.jneumeth.2018.03.020}
 }
 
 @misc{varjo,
@@ -35,11 +37,11 @@
 }
 
 @misc{Unity,
-    title = {Unity},
-    author = {{Unity Technologies}},
-    year = {2005},
-    url = {https://unity.com/},
-    note = {Accessed: May 10, 2023}
+  title = {Unity},
+  author = {{Unity Technologies}},
+  year = {2005},
+  url = {https://unity.com/},
+  note = {Accessed: May 10, 2023}
 }
 
 @misc{matlab,
@@ -59,14 +61,14 @@
 }
 
 @software{kothe2014labstreaminglayer,
-  title = {A software frame- work for synchronizing a large array of data collection and stimulation devices.},
+  title = {A software framework for synchronizing a large array of data collection and stimulation devices.},
   author = {Kothe, Christian and Mandel, Christian},
   url = {https://github.com/sccn/labstreaminglayer},
   note = {Accessed: June 7, 2023}
 }
 
 @article{rama2010eye,
-  title={Eye fixation--related potentials (EFRPs) during object identification},
+  title={Eye fixation-related potentials (EFRPs) during object identification},
   author={R{\"a}m{\"a}, PIA and Baccino, Thierry},
   journal={Visual Neuroscience},
   volume={27},
@@ -74,6 +76,7 @@
   pages={187--192},
   year={2010},
   publisher={Cambridge University Press}
+  doi={10.1017/S0952523810000283}
 }
 
 @article{linden2005p300,
@@ -85,6 +88,7 @@
   pages={563--576},
   year={2005},
   publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+  doi={10.1177/1073858405280524}
 }
 
 @online{zeromq,
@@ -103,18 +107,9 @@
   pages={066052},
   year={2022},
   publisher={IOP Publishing}
+  doi={10.1088/1741-2552/ac4593}
 }
 
-@article{wei2022indexpen,
-  title={IndexPen: Two-Finger Text Input with Millimeter-Wave Radar},
-  author={Wei, Haowen and Li, Ziheng and Galvan, Alexander D and Su, Zhuoran and Zhang, Xiao and Pahlavan, Kaveh and Solovey, Erin T},
-  journal={Proceedings of the ACM on Interactive, Mobile, Wearable and Ubiquitous Technologies},
-  volume={6},
-  number={2},
-  pages={1--39},
-  year={2022},
-  publisher={ACM New York, NY, USA}
-}
 
 @inproceedings{salvucci2000identifying,
   title={Identifying fixations and saccades in eye-tracking protocols},
@@ -122,6 +117,7 @@
   booktitle={Proceedings of the 2000 symposium on Eye tracking research \& applications},
   pages={71--78},
   year={2000}
+  doi={10.1145/355017.355028}
 }
 
 @article{itti1998model,
@@ -133,6 +129,7 @@
   pages={1254--1259},
   year={1998},
   publisher={Ieee}
+  doi={10.1109/34.730558}
 }
 
 @inproceedings{steil2018fixation,
[…]
