### 2023-09-11 #5838 @ApocalyVec
> @ApocalyVec thanks for this submission. I am the AEiC on this track and here to help with the initial steps. I will be looking for a handling editor. For the moment can you check those potentially missing DOIs? ☝️ You can amend your .bib file and rerun (in a clean comment here): `@editorialbot generate pdf`
>
> Followed by: `@editorialbot check references`
>
> Thanks.

Thank you @Kevin-Mattheus-Moerman for noting this. I just updated paper.bib to include the missing DOIs. Maybe it's a good idea to tell the authors, in the submission instructions, to include them in their .bib file. Although the example paper.bib does have DOIs. The BibTeX generated by Google Scholar doesn't include DOI by default.
