### 2023-09-11 804592d
fix syntax in paper.bib
