+++ paper/paper.bib
@@ -6,7 +6,7 @@
   number={1},
   pages={35--53},
   year={2010},
-  publisher={MIT press}
+  publisher={MIT press},
   doi={10.1162/pres.19.1.35}
 }
 
@@ -17,7 +17,7 @@
   volume={303},
   pages={55--67},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.jneumeth.2018.03.020}
 }
 
@@ -75,7 +75,7 @@
   number={5-6},
   pages={187--192},
   year={2010},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
   doi={10.1017/S0952523810000283}
 }
 
@@ -87,7 +87,7 @@
   number={6},
   pages={563--576},
   year={2005},
-  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA},
   doi={10.1177/1073858405280524}
 }
 
@@ -106,7 +106,7 @@
   number={6},
   pages={066052},
   year={2022},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
   doi={10.1088/1741-2552/ac4593}
 }
 
@@ -116,7 +116,7 @@
   author={Salvucci, Dario D and Goldberg, Joseph H},
   booktitle={Proceedings of the 2000 symposium on Eye tracking research \& applications},
   pages={71--78},
-  year={2000}
+  year={2000},
   doi={10.1145/355017.355028}
 }
 
@@ -128,7 +128,7 @@
   number={11},
   pages={1254--1259},
   year={1998},
-  publisher={Ieee}
+  publisher={Ieee},
   doi={10.1109/34.730558}
 }
 
@@ -137,7 +137,7 @@
   author={Steil, Julian and Huang, Michael Xuelin and Bulling, Andreas},
   booktitle={Proceedings of the 2018 ACM Symposium on eye tracking research \& applications},
   pages={1--9},
-  year={2018}
+  year={2018},
   doi={10.1145/3204493.3204538}
 }
 
@@ -149,7 +149,7 @@
   number={7},
   pages={4--4},
   year={2012},
-  publisher={The Association for Research in Vision and Ophthalmology}
+  publisher={The Association for Research in Vision and Ophthalmology},
   doi={10.1167/12.7.4}
 }
 
@@ -161,7 +161,7 @@
   number={1-2},
   pages={8--13},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.jneumeth.2006.11.017}
 }
 
@@ -185,7 +185,7 @@
   journal={bioRxiv},
   pages={2022--12},
   year={2022},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
   doi={10.1101/2022.12.16.520764}
 }
 
@@ -196,7 +196,7 @@
   volume={14},
   pages={900},
   year={2020},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
   doi={10.3389/fnins.2020.00900}
 }
 
@@ -206,7 +206,7 @@
   booktitle={2016 IEEE winter conference on applications of computer vision (WACV)},
   pages={1--10},
   year={2016},
-  organization={IEEE}
+  organization={IEEE},
   doi={10.1109/WACV.2016.7477553}
 }
 
@@ -218,7 +218,7 @@
   number={5},
   pages={056014},
   year={2013},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
   doi={10.1088/1741-2560/10/5/056014}
 }
 
[…]
