+++ paper/paper.bib
@@ -170,7 +170,7 @@
   author = {{Tobii AB}},
   howpublished = {{Tobii AB}},
   year = {2023},
-  url = {{https://www.tobii.com/}}
+  url = {https://www.tobii.com/}
 }
 
 @misc{nirx,
