+++ tests/DummyTest.sh
@@ -1,0 +1,1 @@
+pytest tests/DummyTest.py
