+++ .github/workflows/unit-tests.yml
@@ -1,5 +1,5 @@
-name: unit-test
-run-name: Run unit tests by @${{ github.actor }}
+name: unit-tests
+run-name: Run unit tests after push by @${{ github.actor }}
 on:
   push:
     branches: [ test-actions ]
@@ -24,4 +24,5 @@ jobs:
           /sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
           pip install -r requirements.dev.txt
       - name: Test with pytest
-        run: ./tests/test.sh
+#        run: ./tests/DummyTest.sh
+        run: ./tests/DummyTest.sh
