+++ .github/workflows/unit-tests.yml
@@ -24,4 +24,10 @@ jobs:
           /sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
           pip install -r requirements.dev.txt
       - name: Test with pytest
-        run: ./tests/DummyTest.sh
+
+        run: |
+          ls
+          cd tests
+          ls
+          cd ..
+          ./tests/DummyTest.sh
