+++ .github/workflows/unit-tests.yml
@@ -14,6 +14,7 @@ jobs:
       DISPLAY: ':99.0'
     steps:
       - uses: actions/checkout@v3
+      - uses: tlambert03/setup-qt-libs@v1
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
