### 2023-09-14 87a7aca
add contain function in data buffer
