+++ physiolabxr/scripting/AOIAugmentationScript/AOIAugmentationConfig.py
@@ -19,7 +19,7 @@ image_center_x = 0
 image_center_y = 0
 
 image_shape = np.array([512, 1024], dtype=np.int32)
-attention_patch_shape = np.array([32, 32], dtype=np.int32)
+attention_patch_shape = np.array([16, 32], dtype=np.int32)
 attention_grid_shape = np.array(
     [image_shape[0] // attention_patch_shape[0], image_shape[1] // attention_patch_shape[1]], dtype=np.int32)
 
