### 2023-09-18 #5854 @editorialbot
[REVIEW]: PhysioLabXR: A Python Platform for Real-Time, Multi-modal, Brain–Computer Interfaces and Extended Reality Experiments
**Submitting author:** @ApocalyVec (<a href="http://orcid.org/0000-0001-5187-200X">Ziheng Li</a>)
**Repository:** https://github.com/PhysioLabXR/PhysioLabXR
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.1.1
