### 2023-09-19 4af45b1
update attention generation
