+++ physiolabxr/examples/ZMQExampleOutlet.py
@@ -13,11 +13,11 @@ from pylsl import local_clock
 def main():
     topic = "CamCapture"
     srate = 120
-    port = "5557"
+    port = "5558"
 
     c_channels = 3
-    width = 64
-    height = 64
+    width = 400
+    height = 400
     n_channels = c_channels * width * height
 
     context = zmq.Context()
