+++ physiolabxr/scripting/AOIAugmentationScript/preprocessing/generate_pickle.py
@@ -16,16 +16,16 @@ from eidl.utils.model_utils import get_trained_model, load_image_preprocess
 import cv2
 
 
-class ImageAttentionInfo():
-    def __init__(self):
-        self.image_path = None
-        self.image = None
-        self.image_shape = None
-        self.attention_patch_shape = None
-        self.attention_matrix = None
+class ImageInfo():
+    def __init__(self,image_path, image, image_normalized, attention_patch_shape, attention_matrix, y=None, y_pred=None):
+        self.image_path = image_path
+        self.image = image
+        self.image_normalized = image_normalized
+        self.attention_patch_shape = attention_patch_shape
+        self.attention_matrix = attention_matrix
+        self.y = y
+        self.y_pred = y_pred
 
-        self.y = None
-        self.y_pred = None
 
 
 
@@ -38,7 +38,7 @@ model, image_mean, image_std, image_size, compound_label_encoder = get_trained_m
 ##########################################################################
 
 
-root_dir = r'D:\HaowenWei\PycharmProjects\PhysioLabXR\physiolabxr\scripting\AOIAugmentationScript\data\reports_cleaned\G'
+root_dir = r'D:\HaowenWei\UnityProject\PerceptualAOIAugmentation\Assets\Prefabs\OCTReportImages\Practice'
 image_names = [file for file in os.listdir(root_dir) if file.endswith('.png')]
 
 
@@ -46,81 +46,42 @@ data_dict = {}
 
 y_true = 'G'
 
-for image_name in image_names:
+for index, image_name in enumerate(image_names):
 
     image_path = os.path.join(root_dir, image_name)
 
-    image_meta_info = image_name.split('_')
-
-
-    image_attention_info = ImageAttentionInfo()
     # get the prediction and attention matrix
     image_normalized, image = load_image_preprocess(image_path, image_size, image_mean, image_std) # the normalized image is z normalization
     y_pred, attention_matrix = model(torch.Tensor(image_normalized).unsqueeze(0).to(device),
                                      collapse_attention_matrix=False)
+
     predicted_label = np.array([torch.argmax(y_pred).item()])
     print(f'y_pred: {y_pred}')
     decoded_label = compound_label_encoder.decode(predicted_label)
     print(f'Predicted label: {decoded_label}')
 
-    ##################################
-    # plot the image
-    # if y_true != decoded_label:
     plt.imshow(image.astype(np.uint8))
     plt.title(f'y_true: {[y_true]}, y_pred: {decoded_label}')
     plt.show()
 
-
-
-    # detach the attention matrix
     attention_matrix = attention_matrix.squeeze().cpu().detach().numpy()
 
-    # remove the class token
-    attention_matrix = attention_matrix[1:, 1:]
+    class_token_attention = attention_matrix[0, 1:]
+
+    attention_grid_mask = generate_attention_grid_mask(class_token_attention, attention_patch_shape=AOIAugmentationConfig.attention_patch_shape)
 
 
-    image_attention_info.image_path = image_path
-    image_attention_info.image = image
-    image_attention_info.image_shape = image.shape
-    image_attention_info.attention_patch_shape = attention_matrix.shape
-    image_attention_info.attention_matrix = attention_matrix
-
-    image_attention_info.y = image_meta_info[1]
-    image_attention_info.y_pred = predicted_label
-
-    # data_dict[int(image_meta_info[0])] = image_attention_info
-
-
-
-
-    # post analysis
-    average_attention = np.mean(attention_matrix, axis=1)
-
-    # reshape to attention grid
[…]
