### 2023-09-26 63bead3
add practice image info to git
