### 2023-09-28 7a7d6bd
archive all, switch to bounding box visualization
