+++ physiolabxr/scripting/AOIAugmentationScript/preprocessing/generate_pickle.py
@@ -1,5 +1,8 @@
 import time
 from collections import deque
+
+from eidl.viz.vit_rollout import VITAttentionRollout
+
 from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeData, \
     GazeFilterFixationDetectionIVT, \
     tobii_gaze_on_display_area_to_image_matrix_index, GazeType, gaze_point_on_image_valid
@@ -38,7 +41,7 @@ model, image_mean, image_std, image_size, compound_label_encoder = get_trained_m
 ##########################################################################
 
 
-root_dir = r'D:\HaowenWei\UnityProject\PerceptualAOIAugmentation\Assets\Prefabs\OCTReportImages\Practice'
+root_dir = r'D:\HaowenWei\UnityProject\PerceptualAOIAugmentation\Assets\Prefabs\ExperimentImages\Practice'
 image_names = [file for file in os.listdir(root_dir) if file.endswith('.png')]
 
 
@@ -52,7 +55,9 @@ for index, image_name in enumerate(image_names):
 
     # get the prediction and attention matrix
     image_normalized, image = load_image_preprocess(image_path, image_size, image_mean, image_std) # the normalized image is z normalization
-    y_pred, attention_matrix = model(torch.Tensor(image_normalized).unsqueeze(0).to(device),
+
+    image_tensor = torch.Tensor(image_normalized).unsqueeze(0).to(device)
+    y_pred, attention_matrix = model(image_tensor,
                                      collapse_attention_matrix=False)
 
     predicted_label = np.array([torch.argmax(y_pred).item()])
@@ -62,30 +67,59 @@ for index, image_name in enumerate(image_names):
 
     plt.imshow(image.astype(np.uint8))
     plt.title(f'y_true: {[y_true]}, y_pred: {decoded_label}')
+    plt.colorbar()
     plt.show()
 
-    attention_matrix = attention_matrix.squeeze().cpu().detach().numpy()
+    vit_rollout = VITAttentionRollout(model, device=device, attention_layer_name='attn_drop', head_fusion="mean",
+                                      discard_ratio=0.5)
+    rollout = vit_rollout(depth=model.depth, input_tensor=image_tensor)
 
-    class_token_attention = attention_matrix[0, 1:]
+    rollout_resized = cv2.resize(rollout, dsize=image_size, interpolation=cv2.INTER_LINEAR)
+    rollout_heatmap = cv2.applyColorMap(cv2.cvtColor((rollout_resized * 255).astype(np.uint8), cv2.COLOR_GRAY2BGR),
+                                        cv2.COLORMAP_JET)
+    rollout_heatmap = cv2.cvtColor(rollout_heatmap, cv2.COLOR_BGR2RGB)
+    alpha = 0.2
+    output_image = cv2.addWeighted(image.astype(np.uint8), alpha, rollout_heatmap, 1 - alpha, 0)
 
-    attention_grid = class_token_attention.reshape(AOIAugmentationConfig.attention_grid_shape)
+    fig = plt.figure(figsize=(15, 30), constrained_layout=True)
+    axes = fig.subplots(3, 1)
+    axes[0].imshow(image.astype(np.uint8))  # plot the original image
+    axes[0].axis('off')
+    axes[0].set_title(f'Original image')
 
-    attention_grid_upsample = np.repeat(attention_grid,2, axis=1)
-    plt.imshow(attention_grid_upsample)
+    axes[1].imshow(rollout_heatmap)  # plot the attention rollout
+    axes[1].axis('off')
+    axes[1].set_title(f'Attention rollout')
+
+    axes[2].imshow(output_image)  # plot the attention rollout
+    axes[2].axis('off')
[…]
