+++ physiolabxr/scripting/AOIAugmentationScript/AOIAugmentationScript.py
@@ -29,11 +29,11 @@ class AOIAugmentationScript(RenaScript):
 
         self.device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
 
-        self.gaze_attention_matrix = GazeAttentionMatrixTorch(
-            image_shape=AOIAugmentationConfig.image_shape,
-            attention_patch_shape=AOIAugmentationConfig.attention_patch_shape,
-            device=self.device
-        )
+        # self.gaze_attention_matrix = GazeAttentionMatrixTorch(
+        #     image_shape=AOIAugmentationConfig.image_shape,
+        #     attention_patch_shape=AOIAugmentationConfig.attention_patch_shape,
+        #     device=self.device
+        # )
         # self.gaze_attention_clutter_removal_data_processor = ClutterRemoval(signal_clutter_ratio=0.1)
         # self.gaze_attention_clutter_removal_data_processor.evoke_data_processor()
 
@@ -43,36 +43,36 @@ class AOIAugmentationScript(RenaScript):
         self.ivt_filter = GazeFilterFixationDetectionIVT(angular_speed_threshold_degree=100)
         self.ivt_filter.evoke_data_processor()
 
-        ################################################################################################################
-        static_aoi_augmentation_lsl_outlet_info = StreamInfo(
-            AOIAugmentationConfig.StaticAOIAugmentationStateLSLStreamInfo.StreamName,
-            AOIAugmentationConfig.StaticAOIAugmentationStateLSLStreamInfo.StreamType,
-            AOIAugmentationConfig.StaticAOIAugmentationStateLSLStreamInfo.ChannelNum,
-            AOIAugmentationConfig.StaticAOIAugmentationStateLSLStreamInfo.NominalSamplingRate,
-            channel_format=cf_float32)
-        self.static_aoi_augmentation_lsl_outlet = StreamOutlet(static_aoi_augmentation_lsl_outlet_info)
-
-        ################################################################################################################
-        interactive_aoi_augmentation_lsl_outlet_info = StreamInfo(
-            AOIAugmentationConfig.InteractiveAOIAugmentationStateLSLStreamInfo.StreamName,
-            AOIAugmentationConfig.InteractiveAOIAugmentationStateLSLStreamInfo.StreamType,
-            AOIAugmentationConfig.InteractiveAOIAugmentationStateLSLStreamInfo.ChannelNum,
-            AOIAugmentationConfig.InteractiveAOIAugmentationStateLSLStreamInfo.NominalSamplingRate,
-            channel_format=cf_float32)
-        self.interactive_aoi_augmentation_lsl_outlet = StreamOutlet(interactive_aoi_augmentation_lsl_outlet_info)
-        ################################################################################################################
-
-        ################################################################################################################
-        gaze_attention_map_lsl_outlet_info = StreamInfo(
-            AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.StreamName,
-            AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.StreamType,
-            AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.ChannelNum,
-            AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.NominalSamplingRate,
-            channel_format=cf_float32)
-
-        self.gaze_attention_map_lsl_outlet = StreamOutlet(gaze_attention_map_lsl_outlet_info)  # shape: (1250, 1)
-
-        ################################################################################################################
+        # ################################################################################################################
+        # static_aoi_augmentation_lsl_outlet_info = StreamInfo(
[…]
