+++ physiolabxr/scripting/UnityObjectDetection.py
@@ -4,6 +4,8 @@ import cv2
 import zmq
 import numpy as np
 from physiolabxr.scripting.RenaScript import RenaScript
+from physiolabxr.sub_process.TCPInterface import RenaTCPInterface
+
 
 def get_od_model(config_path, weights_path, input_size):
     net = cv2.dnn_DetectionModel(weights_path, config_path)
@@ -18,8 +20,41 @@ def get_class_names(class_file):
         class_name = f.read().rstrip('\n').split('\n')
     return class_name
 
-def process_received_camera_images(image_data, net, class_names, image_shape, threshold=0.45, nms_threshold=0.2):
+def processDepthImage(depthROI):
+    near = 0.1
+    far = 20.0
+    max16bitval = 65535
+    min16bitval = 0
+    filtered_depthROI = depthROI[depthROI != 0]
+    # scale from 0-65535 to 0-1 value range
+    scale = 1.0 / (max16bitval - min16bitval)
+    compressed = filtered_depthROI * scale
+    # decompress values by 0.25 compression factor
+    decompressed = np.power(compressed, 4)
+    # remove non valid 0 depth values
+    valid_decompressed = decompressed[decompressed != 0]
+    # scale from eye to far rather than near to far (still linear 0-1 range)
+    scaled_eye_far = -(valid_decompressed - 1) / (1 + near / far) + near / far
+    scaled_eye_far = scaled_eye_far[scaled_eye_far != 0]
+
+    # remove noisy outliers (there seem to be higher depth values than there should be)
+    mean = np.mean(scaled_eye_far)
+    standard_deviation = np.std(scaled_eye_far)
+    distance_from_mean = abs(scaled_eye_far - mean)
+    max_deviations = 2
+    not_outlier = distance_from_mean < max_deviations * standard_deviation
+    no_outliers = scaled_eye_far[not_outlier]
+    if len(no_outliers) == 0:
+       return None
+    return np.min(no_outliers), np.max(no_outliers), np.average(no_outliers)
+
+def process_received_camera_images(image_data, image_depth_data, net, class_names, image_shape, threshold=0.45, nms_threshold=0.2):
     color_img = image_data.reshape(image_shape).astype(np.uint8)
+    image_depth = image_depth_data.reshape(image_shape[:2])
+
+    minDepth = []
+    maxDepth = []
+    aveDepth = []
 
     classIds, confs, bbox = net.detect(color_img, confThreshold=threshold)
     bbox = list(bbox)
@@ -38,13 +73,34 @@ def process_received_camera_images(image_data, net, class_names, image_shape, th
         ws.append(int(w))
         hs.append(int(h))
 
+        # Process depth information of bounding box region (min, max, depth)
+        depthROI = image_depth[y:y + h, x:x + w]
+
+        depth_results = processDepthImage(depthROI)
+        if depth_results is None:
+            # warnings.warn(f"no valid depth point was found for class {classNames[class_id - 1]}")
+            continue  # go to the next loop (next detected object)
+        minD, maxD, aveD = depth_results
+        minDepth.append(minD)
+        maxDepth.append(maxD)
+        aveDepth.append(aveD)
+
         # Yolo 2D bb visualization
         detected_classes.append(int(class_id))
         cv2.rectangle(color_img, (x, y), (x + w, h + y), color=(0, 255, 0), thickness=2)
         cv2.putText(color_img, class_names[class_id - 1].upper(),
                     (np.max((0, np.min((image_shape[0], box[0] + 10)))),
                      np.max((0, np.min((image_shape[1], box[1] + 30))))),
-                    cv2.FONT_HERSHEY_COMPLEX, 1, (0, 255, 0), 2)
[…]
