### 2023-09-28 7dc3b69
archive attention plot function TODO: create image meta info
