+++ physiolabxr/scripting/AOIAugmentationScript/preprocessing/generate_attention_images.py
@@ -11,28 +11,25 @@ from eidl.viz.vit_rollout import VITAttentionRollout
 
 
 class ImageInfo():
-    def __init__(self, image_path, image, image_normalized, attention_patch_shape, raw_attention_matrix,
+    def __init__(self,image_path, image, image_normalized, attention_patch_shape, raw_attention_matrix,
                  rollout_attention_matrix, average_self_attention_matrix,
                  y=None, y_pred=None):
+
         self.image_path = image_path
         self.image = image
         self.image_normalized = image_normalized
         self.attention_patch_shape = attention_patch_shape
-        self.attention_grid_shape = (
-        image.shape[0] // attention_patch_shape[0], image.shape[1] // attention_patch_shape[1])
+        self.attention_grid_shape = (image.shape[0] // attention_patch_shape[0], image.shape[1] // attention_patch_shape[1])
         self.raw_attention_matrix = raw_attention_matrix
 
         self.y = y
         self.y_pred = y_pred
 
-
-def overlay_heatmap(original, heatmap, image_size, alpha=.5, interpolation=cv2.INTER_NEAREST, cmap=cv2.COLORMAP_JET,
-                    normalize=False):
+def overlay_heatmap(original, heatmap, image_size, alpha=.5, interpolation=cv2.INTER_NEAREST, cmap=cv2.COLORMAP_JET, normalize=False):
     if normalize:
         heatmap = heatmap / heatmap.max()
     heatmap_processed = cv2.resize(heatmap, dsize=image_size, interpolation=interpolation)
-    heatmap_processed = cv2.applyColorMap(cv2.cvtColor((heatmap_processed * 255).astype(np.uint8), cv2.COLOR_GRAY2BGR),
-                                          cmap)
+    heatmap_processed = cv2.applyColorMap(cv2.cvtColor((heatmap_processed * 255).astype(np.uint8), cv2.COLOR_GRAY2BGR), cmap)
     heatmap_processed = cv2.cvtColor(heatmap_processed, cv2.COLOR_BGR2RGB)
     return cv2.addWeighted(original.astype(np.uint8), alpha, heatmap_processed, 1 - alpha, 0), heatmap_processed
 
@@ -40,7 +37,7 @@ def overlay_heatmap(original, heatmap, image_size, alpha=.5, interpolation=cv2.I
 ##########################################################################
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 model, image_mean, image_std, image_size, compound_label_encoder = get_trained_model(device,
-                                                                                     model_param='num-patch-32_image-size-1024-512')
+                                                                                   model_param='num-patch-32_image-size-1024-512')
 ##########################################################################
 
 
@@ -52,14 +49,15 @@ y_true = 'S'
 
 image_names = [file for file in os.listdir(root_dir) if file.endswith('.png') or file.endswith('.jpg')]
 
+
 data_dict = {}
 
 for index, image_name in enumerate(image_names):
+
     image_path = os.path.join(root_dir, image_name)
 
     # get the prediction and attention matrix
-    image_normalized, image = load_image_preprocess(image_path, image_size, image_mean,
-                                                    image_std)  # the normalized image is z normalization
+    image_normalized, image = load_image_preprocess(image_path, image_size, image_mean, image_std) # the normalized image is z normalization
 
     image_tensor = torch.Tensor(image_normalized).unsqueeze(0).to(device)
     y_pred, attention_matrix = model(image_tensor,
[…]
