### 2023-09-28 2d2f736
update pickle meta file generation
