### 2023-09-29 70e3e0c
Handling event markers and buffering
