+++ physiolabxr/scripting/Examples/PhysioLabXR_Balanceball_Demo/MotorImageryBalanceBall.py
@@ -1,0 +1,171 @@
+from enum import Enum
+
+import mne
+import numpy as np
+from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
+from sklearn.model_selection import ShuffleSplit, cross_val_score
+from sklearn.pipeline import Pipeline
+
+from mne import Epochs, events_from_annotations, pick_types, create_info
+from mne.channels import make_standard_montage
+from mne.datasets import eegbci
+from mne.decoding import CSP
+from mne.io import concatenate_raws, read_raw_edf
+
+from physiolabxr.scripting.RenaScript import RenaScript
+from physiolabxr.scripting.physio.epochs import get_event_locked_data
+from physiolabxr.utils.buffers import DataBuffer
+
+event_marker_stream_name = 'EventMarker_BallGame'
+class GameStates(Enum):
+    idle = 'idle'
+    train = 'train'
+    fit = 'fit'
+    eval = 'eval'
+    
+class Events(Enum):
+    train_start = 1
+    left_trial = 2
+    right_trial = 3
+    eval_start = 6
+
+class CSPDecoder:
+
+    def __init__(self, n_components=4):
+        self.n_components = n_components
+        self.csp = CSP(n_components=n_components, reg=None, log=True, norm_trace=False)
+        self.lda = LinearDiscriminantAnalysis()
+        self.clf = Pipeline([('CSP', self.csp), ('LDA', self.lda)])
+
+    def fit(self, X, y):
+        X = self.csp.fit_transform(X, y)
+        # fit classifier
+        self.lda.fit(X, y)
+
+    def csp_transfomr(self, X):
+        return self.csp.transform(X)
+
+    def transform(self, X):
+        X = self.csp.transform(X)
+        return self.lda.transform(X)
+
+def sigmoid(z):
+    return 1/(1 + np.exp(-z))
+
+class MotorImageryBalanceBall(RenaScript):
+    def __init__(self, *args, **kwargs):
+        """
+        Please do not edit this function
+        """
+        super().__init__(*args, **kwargs)
+
+    # Start will be called once when the run button is hit.
+    def init(self):
+        # self.train_data_buffer = DataBuffer()
+        # self.eval_data_buffer = DataBuffer()
+        self.cur_state = 'idle'
+        self.transition_markers = [Events.train_start.value, -Events.train_start.value, Events.eval_start.value, -Events.eval_start.value]
+        self.eeg_channels = ["F3", "Fz", "F4", "C3", "Cz", "C4", "P3", "P4", ]
+        self.decoder_tmin = 2.
+        self.decoder_tmax = 5.
+        self.srate = 128
+        self.decode_t_len = int((self.decoder_tmax - self.decoder_tmin) * self.srate)
+        self.label_mapping = {2: 0, 3: 1}
+        self.decoder = None
+
+        # loop is called <Run Frequency> times per second
+    def loop(self):
+
+        if event_marker_stream_name not in self.inputs.keys(): # or  #EVENT_MARKER_CHANNEL_NAME not in self.inputs.keys():
+            # print('Event marker stream not found')
+            return
+
+        self.process_event_markers()
+        if self.cur_state == GameStates.train:
+            pass
+            # keep collecting data
+            # print("In training")
+        elif self.cur_state == GameStates.eval:
+            # make predictions
+            self.decode()
+            pass
+
+    # cleanup is called when the stop button is hit
+    def cleanup(self):
+        print('Cleanup function is called')
+
+    def process_event_markers(self):
+        if event_marker_stream_name in self.inputs.keys() and len(np.intersect1d(self.inputs[event_marker_stream_name][0], self.transition_markers)) > 0:
+            last_processed_marker_index = None
[…]
