### 2023-09-30 5f7b526
contour generation done. TODO: grid search for the threshold value. Third order standard deviation
