### 2023-09-30 fd56396
update visualization method
