+++ physiolabxr/scripting/AOIAugmentationScript/preprocessing/generate_attention_images.py
@@ -33,29 +33,53 @@ class ImageInfo():
 
 def overlay_heatmap(original, heatmap, image_size, alpha=.5, interpolation=cv2.INTER_NEAREST, cmap=cv2.COLORMAP_JET, normalize=False):
     if normalize:
+        heatmap = heatmap / heatmap.max()
         # min max normalization
-        heatmap = (heatmap - heatmap.min()) / (heatmap.max() - heatmap.min())
+        # heatmap = (heatmap - heatmap.min()) / (heatmap.max() - heatmap.min())
     heatmap_processed = cv2.resize(heatmap, dsize=image_size, interpolation=interpolation)
     heatmap_processed = cv2.applyColorMap(cv2.cvtColor((heatmap_processed * 255).astype(np.uint8), cv2.COLOR_GRAY2BGR), cmap)
     heatmap_processed = cv2.cvtColor(heatmap_processed, cv2.COLOR_BGR2RGB)
     return cv2.addWeighted(original.astype(np.uint8), alpha, heatmap_processed, 1 - alpha, 0), heatmap_processed
 
 
+def overlay_contour(original, heatmap, image_size, threshold=0.5, interpolation=cv2.INTER_NEAREST, color=(0, 0, 255)):
+    # resize the heatmap to the original image size
+    heatmap_processed = cv2.resize(heatmap, dsize=image_size, interpolation=interpolation)
+    ret, thresh = cv2.threshold(heatmap_processed, threshold, 1, cv2.THRESH_BINARY)
+    contours, hierarchy = cv2.findContours(thresh.astype(np.uint8), cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
+    countour_image = cv2.drawContours(original.astype(np.uint8), contours, -1, color, 5)
 
-target_dir = r'D:\HaowenWei\PycharmProjects\PhysioLabXR\physiolabxr\scripting\AOIAugmentationScript\data\report_cleand_attention_vis\S'
+    return countour_image, thresh, contours
+
+
+
+
+target_dir = r'D:\HaowenWei\PycharmProjects\PhysioLabXR\physiolabxr\scripting\AOIAugmentationScript\data\report_cleaned_attention_vis'
+
+
 
 data_dict_path = r'report_cleaned_image_info.pkl'
 with open(data_dict_path, 'rb') as f:
     data_dict = pickle.load(f)
 
 for class_name in data_dict:
+    # if os.path.exists(os.path.join(target_dir, class_name)):
+    #     os.remove(os.path.join(target_dir, class_name))
+
+    os.mkdir(os.path.join(target_dir, class_name))
+
     class_dict = data_dict[class_name]
     for image_name in class_dict:
         image_info = class_dict[image_name]
 
         image = image_info.image_to_model
+        image_normalized = image_info.image_to_model_normalized
+        image_normalized = np.moveaxis(image_normalized, 0, -1)
+
         image_shape = image_info.model_image_shape
         rollout = image_info.rollout_attention_matrix
+        attention_matrix_self_attention = image_info.average_self_attention_matrix
+
         y_true = image_info.y_true
         y_pred = image_info.y_pred
 
@@ -64,37 +88,71 @@ for class_name in data_dict:
         rollout_overlap, rollout_heatmap = overlay_heatmap(image, rollout, (image_shape[1], image_shape[0]),
                                                            alpha=0.5,
                                                            interpolation=cv2.INTER_NEAREST, cmap=cv2.COLORMAP_VIRIDIS)
-        attention_matrix_self_attention = image_info.average_self_attention_matrix
+
         self_attention_overlap, self_attention_heatmap = overlay_heatmap(image, attention_matrix_self_attention,
[…]
