### 2023-10-02 eadc7dc
run contour detection in real-time, compile contour hierarchy in lvt format TODO: send lvt to Unity and plot
