+++ physiolabxr/scripting/AOIAugmentationScript/AOIAugmentationUtils.py
@@ -5,9 +5,133 @@ import numpy as np
 import torch
 import torch.nn.functional as F
 import cv2
+import pickle
 
-from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationConfig import IMAGE_FORMAT
 
+# from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationConfig import IMAGE_FORMAT
+
+
+class ImageInfo():
+    def __init__(self, image_path, original_image,
+                 image_to_model, image_to_model_normalized, model_image_shape,
+                 patch_shape, attention_grid_shape,
+                 raw_attention_matrix,
+                 rollout_attention_matrix, average_self_attention_matrix,
+                 y_true=None, y_pred=None):
+        self.image_path = image_path
+        self.original_image = original_image
+        self.image_to_model = image_to_model
+        self.image_to_model_normalized = image_to_model_normalized
+        self.model_image_shape = model_image_shape
+        self.patch_shape = patch_shape
+        self.attention_grid_shape = attention_grid_shape
+        self.raw_attention_matrix = raw_attention_matrix
+        self.rollout_attention_matrix = rollout_attention_matrix
+        self.average_self_attention_matrix = average_self_attention_matrix
+        self.y_true = y_true
+        self.y_pred = y_pred
+
+
+        # self.image_on_screen_shape = None
+        # self.contours: ContourInfo
+
+
+
+
+# class ContourInfo():
+#     def __init__(self, contours, hierarchy):
+#         self.contours = contours
+#         self.hierarchy = hierarchy
+
+
+
+
+
+
+def get_report_cleaned_image_info_dict(file_path, merge_dict=True):
+    with open(file_path, 'rb') as f:
+        data_dict = pickle.load(f)
+
+    if merge_dict:
+        data_dict_merged = {}
+        for key in data_dict.keys():
+            data_dict_merged.update(data_dict[key])
+        return data_dict_merged
+    else:
+        return data_dict
+
+def contours_to_lvt(contours, hierarchy, max_length=2048):
+    '''
+    Convert contours to lvt format.
+    :param contours:
+    :param max_length:
+    :return:
+    '''
+    # [contour_num, contour
+
+    contours_lvt = []
+    contours_lvt.append(len(contours))
+    overflow = False
+
+    for contour_index, contour in enumerate(contours):
+        contour_lvt = []
+        contour_hierarchy = hierarchy[0][contour_index].tolist()
+        hierarchy_info_length = len(contour_hierarchy)
+
+        contour_vertices_num = len(contour)
+        contour_points = contour.flatten().tolist()
+
+        contour_lvt.append(contour_index)
+        contour_lvt.append(hierarchy_info_length)
+        contour_lvt+=contour_hierarchy
+        contour_lvt.append(contour_vertices_num)
+        contour_lvt+=contour_points
+
+        contours_lvt+=contour_lvt
+
+        if len(contours_lvt) > max_length:
+            overflow = True
+            break
+
+
+    if not overflow:
+        # pad with zeros
+        contours_lvt += [0] * (max_length - len(contours_lvt))
+
+    return contours_lvt
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+def get_image_on_screen_shape(original_image_width, original_image_height, image_width, image_height, keep_aspect_ratio=True):
+    '''
+    Get the image shape on screen.
+    :param original_image_width:
+    :param original_image_height:
+    :param image_width:
+    :param image_height:
+    :param keep_aspect_ratio:
+    :return: int
+    '''
+
+    aspect_ratio = original_image_width / original_image_height
+
+    if (image_width / image_height > aspect_ratio):
[…]
