+++ physiolabxr/scripting/AOIAugmentationScript/AOIAugmentationScript.py
@@ -1,7 +1,16 @@
 import time
 from collections import deque
+import cv2
+import numpy as np
+import time
+import os
+import pickle
+import sys
+import matplotlib.pyplot as plt
+import torch
 
 from pylsl import StreamInfo, StreamOutlet, cf_float32
+
 
 from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeData, GazeFilterFixationDetectionIVT, \
     tobii_gaze_on_display_area_to_image_matrix_index, GazeType, gaze_point_on_image_valid
@@ -10,6 +19,7 @@ from physiolabxr.scripting.AOIAugmentationScript import AOIAugmentationConfig
 from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationUtils import *
 from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationConfig import EventMarkerLSLStreamInfo, GazeDataLSLStreamInfo
 import torch
+
 
 
 class AOIAugmentationScript(RenaScript):
@@ -25,43 +35,24 @@ class AOIAugmentationScript(RenaScript):
         self.currentBlock: AOIAugmentationConfig.ExperimentBlock = \
             AOIAugmentationConfig.ExperimentBlock.StartBlock
 
-        self.currentReportLabel: int = -1
-
         self.device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
 
-        # self.gaze_attention_matrix = GazeAttentionMatrixTorch(
-        #     image_shape=AOIAugmentationConfig.image_shape,
-        #     attention_patch_shape=AOIAugmentationConfig.attention_patch_shape,
-        #     device=self.device
-        # )
-        # self.gaze_attention_clutter_removal_data_processor = ClutterRemoval(signal_clutter_ratio=0.1)
-        # self.gaze_attention_clutter_removal_data_processor.evoke_data_processor()
+        self.current_image_name = None
+        self.current_image_info= None
 
         self.vit_attention_matrix = ViTAttentionMatrix()
+
+        self.process_gaze_data_time_buffer = deque(maxlen=1000)
+
+        self.report_cleaned_image_info_dict = get_report_cleaned_image_info_dict(AOIAugmentationConfig.ReportCleanedImageInfoFilePath, merge_dict=True)
+
+        print("Experiment started")
         # self.vit_attention_matrix.generate_random_attention_matrix(patch_num=1250)
 
-        self.ivt_filter = GazeFilterFixationDetectionIVT(angular_speed_threshold_degree=100)
-        self.ivt_filter.evoke_data_processor()
+        # self.ivt_filter = GazeFilterFixationDetectionIVT(angular_speed_threshold_degree=100)
+        # self.ivt_filter.evoke_data_processor()
 
-        # ################################################################################################################
-        # static_aoi_augmentation_lsl_outlet_info = StreamInfo(
-        #     AOIAugmentationConfig.StaticAOIAugmentationStateLSLStreamInfo.StreamName,
-        #     AOIAugmentationConfig.StaticAOIAugmentationStateLSLStreamInfo.StreamType,
-        #     AOIAugmentationConfig.StaticAOIAugmentationStateLSLStreamInfo.ChannelNum,
-        #     AOIAugmentationConfig.StaticAOIAugmentationStateLSLStreamInfo.NominalSamplingRate,
-        #     channel_format=cf_float32)
-        # self.static_aoi_augmentation_lsl_outlet = StreamOutlet(static_aoi_augmentation_lsl_outlet_info)
-        #
-        # ################################################################################################################
[…]
