### 2023-10-02 bdb56a3
sent the contour lvt to AOIAugmentationAttentionContourStream
