### 2023-10-04 b0b87d6
finish static AOI Augmentation overlay
