### 2023-10-04 99a4422
finish static AOI Augmentation overlay
