### 2023-10-05 #5854 @lucask07
Hi @ApocalyVec, thanks for this nice work. The paper has a few references to 'real-time'. I consider this a 'performance' claim as per the checkbox above. Could you please elaborate on the paper in terms of what is meant by real-time (milliseconds, microseconds as in an RTOS, etc.)? Further what has been done to verify the real-time performance?

Second, the paper mentions closed-loop control: lines 26-27 "often involving closed-loop interaction systems". Has the software been set up to mediate a closed-loop configuration? If so what sensors and actuators were used and what latency did the software support?

I post these comments here, rather than as an issue on the source repository, since each is very specific to the JOSS paper.
