### 2023-10-05 #5854 @lucask07
@ApocalyVec a few paper grammar notes:

1. Line 87 extra space
2. Line 142 stray period after maximizing (?)
3. Line 151 to 154: unclear / awkward sentence
