### 2023-10-05 #5854 @lucask07
@ApocalyVec

Could you please walk us through the checkbox for "**Human and animal research**: If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item."

I see an example dataset such as [Eye fixation data](https://physiolabxrdocs.readthedocs.io/en/latest/FixationDetection.html) and am not sure if this data is relevant to the policy and if it is relevant if the research complies with the JOSS policy.

Fig. 2 of the manuscript also seems to have human subject data. Please help us understand need for compliance here as well.
