+++ physiolabxr/scripting/AOIAugmentationScript/AOIAugmentationScript.py
@@ -36,16 +36,18 @@ class AOIAugmentationScript(RenaScript):
             AOIAugmentationConfig.ExperimentBlock.StartBlock
 
         self.device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
+        self.gaze_attention_matrix = GazeAttentionMatrix(device=self.device)
 
         self.current_image_name = None
         self.current_image_info= None
 
-        self.vit_attention_matrix = ViTAttentionMatrix()
+        # self.vit_attention_matrix = ViTAttentionMatrix()
+
+        self.ivt_filter = GazeFilterFixationDetectionIVT(angular_speed_threshold_degree=100)
 
         self.process_gaze_data_time_buffer = deque(maxlen=1000)
 
         self.report_cleaned_image_info_dict = get_report_cleaned_image_info_dict(AOIAugmentationConfig.ReportCleanedImageInfoFilePath, merge_dict=True)
-
         print("Experiment started")
         # self.vit_attention_matrix.generate_random_attention_matrix(patch_num=1250)
 
@@ -54,14 +56,14 @@ class AOIAugmentationScript(RenaScript):
 
 
         # ################################################################################################################
-        # gaze_attention_map_lsl_outlet_info = StreamInfo(
-        #     AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.StreamName,
-        #     AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.StreamType,
-        #     AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.ChannelNum,
-        #     AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.NominalSamplingRate,
-        #     channel_format=cf_float32)
-        #
-        # self.gaze_attention_map_lsl_outlet = StreamOutlet(gaze_attention_map_lsl_outlet_info)  # shape: (1250, 1)
+        gaze_attention_map_lsl_outlet_info = StreamInfo(
+            AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.StreamName,
+            AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.StreamType,
+            AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.ChannelNum,
+            AOIAugmentationConfig.AOIAugmentationGazeAttentionMapLSLStreamInfo.NominalSamplingRate,
+            channel_format=cf_float32)
+
+        self.gaze_attention_map_lsl_outlet = StreamOutlet(gaze_attention_map_lsl_outlet_info)  # shape: (1024, 1)
         #
         # ################################################################################################################
 
@@ -143,6 +145,26 @@ class AOIAugmentationScript(RenaScript):
                     if self.current_image_name in self.report_cleaned_image_info_dict.keys():
                         current_image_info_dict = self.report_cleaned_image_info_dict[self.current_image_name]
                         self.current_image_info = ImageInfo(**current_image_info_dict)
+
+                        self.gaze_attention_matrix.set_image_shape(self.current_image_info.model_image_shape)
+                        self.gaze_attention_matrix.set_attention_patch_shape(self.current_image_info.patch_shape)
+
+                        # get image on screen pixel shape
+                        image_on_screen_shape = get_image_on_screen_shape(
+                            original_image_width=self.current_image_info.original_image.shape[1],
+                            original_image_height=self.current_image_info.original_image.shape[0],
+                            image_width=AOIAugmentationConfig.image_on_screen_width,
+                            image_height=AOIAugmentationConfig.image_on_screen_height,
+                        )
+
[…]
