### 2023-10-12 80c2913
finish AOI Interaction mode script
