+++ physiolabxr/scripting/AOIAugmentationScript/AOIAugmentationScript.py
@@ -14,6 +14,7 @@ from pylsl import StreamInfo, StreamOutlet, cf_float32
 
 from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeData, GazeFilterFixationDetectionIVT, \
     tobii_gaze_on_display_area_to_image_matrix_index, GazeType, gaze_point_on_image_valid
+from physiolabxr.scripting.AOIAugmentationScript.IntegrateAttention import integrate_attention
 from physiolabxr.scripting.RenaScript import RenaScript
 from physiolabxr.scripting.AOIAugmentationScript import AOIAugmentationConfig
 from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationUtils import *
@@ -35,11 +36,11 @@ class AOIAugmentationScript(RenaScript):
         self.currentBlock: AOIAugmentationConfig.ExperimentBlock = \
             AOIAugmentationConfig.ExperimentBlock.StartBlock
 
-        self.device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
-        self.gaze_attention_matrix = GazeAttentionMatrix(device=self.device)
+        self.device = torch.device('cpu')
+
 
         self.current_image_name = None
-        self.current_image_info= None
+
 
         # self.vit_attention_matrix = ViTAttentionMatrix()
 
@@ -48,6 +49,17 @@ class AOIAugmentationScript(RenaScript):
         self.process_gaze_data_time_buffer = deque(maxlen=1000)
 
         self.report_cleaned_image_info_dict = get_report_cleaned_image_info_dict(AOIAugmentationConfig.ReportCleanedImageInfoFilePath, merge_dict=True)
+
+        self.gaze_attention_matrix = GazeAttentionMatrix(device=self.device)
+
+        # prevent null pointer in if statement been compiled before run
+        self.gaze_attention_matrix.set_image_shape(np.array([512,1024]))
+        self.gaze_attention_matrix.set_attention_patch_shape(np.array([16,32]))
+        test = self.gaze_attention_matrix.get_gaze_attention_grid_map(flatten=False)
+
+        self.current_image_info = ImageInfo()
+        self.update_cue_now = False
+
         print("Experiment started")
         # self.vit_attention_matrix.generate_random_attention_matrix(patch_num=1250)
 
@@ -79,7 +91,7 @@ class AOIAugmentationScript(RenaScript):
         self.aoi_augmentation_attention_contour_lsl_outlet = StreamOutlet(aoi_augmentation_attention_contour_lsl_outlet_info)  # shape: (1024,)
         # ################################################################################################################
 
-        self.cur_attention_human = None
+        # self.cur_attention_human = None
         # Start will be called once when the run button is hit.
 
 
@@ -94,9 +106,9 @@ class AOIAugmentationScript(RenaScript):
         if (EventMarkerLSLStreamInfo.StreamName not in self.inputs.keys()) or (
                 GazeDataLSLStreamInfo.StreamName not in self.inputs.keys()):  # or GazeDataLSLOutlet.StreamName not in self.inputs.keys():
             return
-
-        self.state_shift()
-
+        # print("process event marker call start")
+        self.process_event_markers()
+        # print("process event marker call complete")
 
         if self.currentExperimentState == AOIAugmentationConfig.ExperimentState.NoAOIAugmentationState:
             self.no_aoi_augmentation_state_callback()
@@ -108,12 +120,13 @@ class AOIAugmentationScript(RenaScript):
     def cleanup(self):
         print('Cleanup function is called')
 
-    def state_shift(self):
+    def process_event_markers(self):
[…]
