+++ .github/workflows/unit-tests.yml
@@ -25,6 +25,7 @@ jobs:
           sudo apt install libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 x11-utils libegl1
           /sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
           sudo apt install qt6-base-dev
+          pip install wheel  
           pip install numpy==1.11.*
           pip install -r requirements.dev.txt
       - name: Test with pytest
