+++ .github/workflows/unit-tests.yml
@@ -26,7 +26,6 @@ jobs:
           /sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
           sudo apt install qt6-base-dev
           pip install wheel  
-          pip install numpy==1.11.*
           pip install -r requirements.dev.txt
       - name: Test with pytest
         run: |
