### 2023-10-18 7e64954
update screen config/setting
