+++ physiolabxr/scripting/AOIAugmentationScript/AOIAugmentationConfig.py
@@ -7,14 +7,14 @@ import pickle
 
 
 
-# ReportCleanedImageInfoFilePath = r'D:\HaowenWei\PycharmProjects\PhysioLabXR\physiolabxr\scripting\AOIAugmentationScript\data\experiment_data\report_cleaned_image_info.pkl'
-ReportCleanedImageInfoFilePath = r'D:\HaowenWei\Rena\PhysioLabXR\physiolabxr\scripting\AOIAugmentationScript\data\experiment_data\report_cleaned_image_info.pkl'
+ReportCleanedImageInfoFilePath = r'D:\HaowenWei\PycharmProjects\PhysioLabXR\physiolabxr\scripting\AOIAugmentationScript\data\experiment_data\report_cleaned_image_info.pkl'
+#ReportCleanedImageInfoFilePath = r'D:\HaowenWei\Rena\PhysioLabXR\physiolabxr\scripting\AOIAugmentationScript\data\experiment_data\report_cleaned_image_info.pkl'
 
 screen_width = 1920
 screen_height = 1080
 
-image_on_screen_width = 2000
-image_on_screen_height = 1000
+image_on_screen_width = 1800
+image_on_screen_height = 900
 
 image_center_x = 0
 image_center_y = 0
