+++ physiolabxr/presets/PlotConfig.py
@@ -31,7 +31,7 @@ class ImageConfig(metaclass=SubPreset):
 
     width: int = 0
     height: int = 0
-    rotation_degree: int = 0
+    rotation_clockwise_degree: int = 0
     scaling_percentage: float = 100
 
     cmap: Cmap = Cmap.VIRIDIS
