+++ physiolabxr/presets/PlotConfig.py
@@ -31,6 +31,7 @@ class ImageConfig(metaclass=SubPreset):
 
     width: int = 0
     height: int = 0
+    rotation_clockwise_degree: int = 0
     scaling_percentage: float = 100
 
     cmap: Cmap = Cmap.VIRIDIS
