### 2023-10-19 6ae0137
fix image format channel first/channel last reversed bug, finish image rotation. it works for rgb,bgr, and pixel map. feature tested user can select rotate clockwise 0,90,180.270 degrees
fix image rotation combobox not set bug
