### 2023-10-20 #5854 @mstimberg
Hi @ApocalyVec, did you have a chance to look at @lucask07 earlier comments? And could you please comment on whether the eye fixation data set is original, otherwise unpublished data that has been collected by the authors of this manuscript.
