### 2023-10-20 f7f646f
adding 'pip install pytest' in the yml file
