### 2023-10-21 9827c5f
make buffer reshape when frame number of channels mismatch what's in the buffer
