### 2023-10-21 7dd3931
starting to write object detection from webcam
