+++ physiolabxr/scripting/Examples/ObjectDetection.py
@@ -5,7 +5,7 @@ import zmq
 import numpy as np
 from physiolabxr.scripting.RenaScript import RenaScript
 
-
+# get the object detection model
 def get_od_model(config_path, weights_path, input_size):
     net = cv2.dnn_DetectionModel(weights_path, config_path)
     net.setInputSize(input_size)
@@ -14,24 +14,26 @@ def get_od_model(config_path, weights_path, input_size):
     net.setInputSwapRB(True)
     return net
 
-
+# get the class names from the coco.names file for the object detection model
 def get_class_names(class_file):
     with open(class_file, 'rt') as f:
         class_name = f.read().rstrip('\n').split('\n')
     return class_name
 
 
-
+# process the received camera images
 def process_received_camera_images(image_data, net, class_names, image_shape, threshold=0.45, nms_threshold=0.2):
-    color_img = image_data.reshape(image_shape).astype(np.uint8)
+    color_img = image_data.reshape(image_shape).astype(np.uint8) # reshape the image data to the image shape
+    color_img = cv2.rotate(color_img, cv2.ROTATE_90_COUNTERCLOCKWISE) # rotate the image 90 degrees counter clockwise because the cv2 has a different origin
 
-    classIds, confs, bbox = net.detect(color_img, confThreshold=threshold)
+    classIds, confs, bbox = net.detect(color_img, confThreshold=threshold) # get the bounding boxes, confidence, and class ids
     bbox = list(bbox)
     confs = list(np.array(confs).reshape(1, -1)[0])
     confs = list(map(float, confs))
 
-    indices = cv2.dnn.NMSBoxes(bbox, confs, threshold, nms_threshold)
+    indices = cv2.dnn.NMSBoxes(bbox, confs, threshold, nms_threshold) # get the indices of the bounding boxes
     detected_classes, xs, ys, ws, hs = list(), list(), list(), list(), list()
+
     for i in indices:
         class_id = classIds[i][0] if type(classIds[i]) is list or type(classIds[i]) is np.ndarray else classIds[i]
         i = i[0] if type(i) is list or type(i) is np.ndarray else i
@@ -50,6 +52,9 @@ def process_received_camera_images(image_data, net, class_names, image_shape, th
                      np.max((0, np.min((image_shape[1], box[1] + 30))))),
                     cv2.FONT_HERSHEY_COMPLEX, 1, (0, 255, 0), 2)
 
+    color_img = cv2.rotate(color_img, cv2.ROTATE_90_CLOCKWISE) # rotate the image back to its original orientation
+
+    # return the detected classes, the positions, and the image with bounding boxes
     return {
         'classIDs': detected_classes,
         'xs': xs,
@@ -58,7 +63,7 @@ def process_received_camera_images(image_data, net, class_names, image_shape, th
         'hs': hs,
     }, color_img
 
-class BaseRenaScript(RenaScript):
+class ObjectDetectionScript(RenaScript):
     def __init__(self, *args, **kwargs):
         """
         Please do not edit this function
@@ -66,7 +71,7 @@ class BaseRenaScript(RenaScript):
         super().__init__(*args, **kwargs)
         config_path = 'physiolabxr/examples/ObjectDetection/ssd_mobilenet_v3_large_coco_2020_01_14.pbtxt'
         weights_path = 'physiolabxr/examples/ObjectDetection/frozen_inference_graph.pb'
-        self.image_shape = (400, 400, 3)
+        self.image_shape = (640, 480, 3)
         self.ob_model = get_od_model(config_path, weights_path, input_size=self.image_shape[:2])
         self.class_names = get_class_names('physiolabxr/examples/ObjectDetection/coco.names')
 
@@ -77,13 +82,11 @@ class BaseRenaScript(RenaScript):
 
[…]
