### 2023-10-26 ce8e704
Update ObjectDetection.py
fixed issue of the webcam having to be called Camera 0. Now it can be any Camera that is in the input
