### 2023-10-26 1defef7
added json file and can now take small pictures of objects intersected with gaze
added json file and can now take small pictures of objects intersected with gaze
