+++ physiolabxr/scripting/Examples/TargetIdentification.py
@@ -1,5 +1,7 @@
+import warnings
 from enum import Enum
 
+import cv2
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -14,7 +16,9 @@ event_marker_block_id_index = 1
 item_marker_dtn_index = 0
 item_marker_gaze_intersected_index = 5
 n_chan_per_item = 10
-subimage_size = 128, 128
+subimage_size = 64, 64
+# subimage_size = 128, 128
+# capture_size = 1920, 1080
 capture_size = 720, 480
 
 class Events(Enum):
@@ -89,6 +93,8 @@ class TargetIdentification(RenaScript):
                 found_event_index = event_indices[0]
 
                 self.next_state = States.idle
+                self.get_gaze_intersected_images()
+
                 print(f"found block end, entering {self.next_state}, the block just ended is {self.cur_block_id}")
 
             if found_event_index is not None:
@@ -99,6 +105,7 @@ class TargetIdentification(RenaScript):
         return np.squeeze(np.argwhere(block_markers == event_value), axis=1)
 
     def get_gaze_intersected_images(self):
+
         captures = self.inputs[capture_stream_name][0]
         capture_times = self.inputs[capture_stream_name][1]
 
@@ -114,6 +121,9 @@ class TargetIdentification(RenaScript):
         gaze_onsets = np.argwhere(item_gaze_intersects == 1)  # first column is the item index, second column is the time index
         gaze_offsets = np.argwhere(item_gaze_intersects == -1)
 
+        if len(gaze_onsets) == 0:
+            warnings.warn("get_gaze_intersected_images: no gaze intersect found, returning")
+            return
         gaze_dtn = item_dtns[gaze_onsets[:, 0]]
         gaze_times = item_markers_times[gaze_onsets[:, 1]]
 
@@ -126,12 +136,22 @@ class TargetIdentification(RenaScript):
 
         # get the camera image
         capture_indices = [np.argmin(np.abs(capture_times - t)) for t in gaze_times]
-        captures = captures[:, capture_indices].reshape(capture_size[0], capture_size[1], 3, -1)  # three for the color channels
-        captures = np.transpose(captures, (3, 0, 1, 2))
+        captures = captures[:, capture_indices].reshape(capture_size[1], capture_size[0], 3, -1)  # three for the color channels
+        captures = np.moveaxis(captures, -1, 0)
+        captures = [cv2.flip(x, 0) for x in captures]
+        gaze_pos[:, 1] = capture_size[1] - gaze_pos[:, 1]
+
+        for i in range(len(captures)):
+            plt.imshow(captures[i])
+            plt.scatter(gaze_pos[i][0], gaze_pos[i][1], color='red', marker='x', linewidths=0.1)
+            plt.show()
+
         # grab the subimages
         for i, (g_pos, cap) in enumerate(zip(gaze_pos, captures)):
-            subimage = cap[int(g_pos[0] - subimage_size[0]/2):int(g_pos[0] + subimage_size[0]/2),
-                                  int(g_pos[1] - subimage_size[1]/2):int(g_pos[1] + subimage_size[1]/2), :]
+            subimage = cap[int(g_pos[1] - subimage_size[0]/2):int(g_pos[1] + subimage_size[0]/2),
+                                  int(g_pos[0] - subimage_size[1]/2):int(g_pos[0] + subimage_size[1]/2), :]
+            # subimage = cap[int(g_pos[0]):int(g_pos[0] + subimage_size[0]),
+            #            int(g_pos[1] - subimage_size[1] / 2):int(g_pos[1] + subimage_size[1] / 2), :]
             self.images.append(subimage)
             plt.imsave(f'{i}.png', subimage)
 
