+++ tests/test_utils.py
@@ -10,6 +10,7 @@ from collections import defaultdict
 from multiprocessing import Process
 from typing import Union, Iterable, List
 
+import matplotlib.pyplot as plt
 import numpy as np
 from PyQt6 import QtCore, QtWidgets
 from PyQt6.QtCore import Qt
@@ -49,6 +50,10 @@ def streams_are_available(app, test_stream_names: List[str]):
         assert app.stream_widgets[ts_name].is_stream_available
 
 def stream_is_unavailable(app_main_window, stream_name):
+    # if type is zmq just, return
+    from physiolabxr.presets.Presets import Presets
+    if Presets().stream_presets[stream_name].preset_type == PresetType.ZMQ:
+        return True
     assert not app_main_window.stream_widgets[stream_name].is_stream_available
 
 def handle_custom_dialog_ok(qtbot, patience_second=0, click_delay_second=0):
@@ -129,7 +134,7 @@ class ContextBot:
     def cleanup(self):
         pass
 
-    def create_add_start_stream(self, stream_name: str, num_channels: int, srate:int, preset_type: PresetType = PresetType.LSL):
+    def create_add_start_stream(self, stream_name: str, num_channels: int, srate: int, preset_type: PresetType = PresetType.LSL, **kwargs):
         """
         start a stream as a separate process, add it to the app's streams, and start it once it becomes
         available
@@ -138,12 +143,19 @@ class ContextBot:
         """
         if stream_name in self.send_data_processes.keys():
             raise ValueError(f"Stream name {stream_name} is in keys for send_data_processes")
-        p = Process(target=LSLTestStream, args=(stream_name, num_channels, srate))
-        p.start()
-        self.send_data_processes[stream_name] = p
+        if preset_type == PresetType.LSL:
+            p = Process(target=LSLTestStream, args=(stream_name, num_channels, srate))
+            p.start()
+            self.send_data_processes[stream_name] = p
+        elif preset_type == PresetType.ZMQ:
+            kwargs['port'] = self.create_zmq_stream(stream_name, num_channels=num_channels, srate=srate, **kwargs)
+        else:
+            raise ValueError(f"ContextBot.create_add_start_stream(): preset_type {preset_type} not supported")
+
         # check the preset hasn't already been added
-        if stream_name not in self.app.presets.keys():
-            self.app.create_preset(stream_name, preset_type, num_channels=num_channels, nominal_sample_rate=srate)
+        from physiolabxr.presets.Presets import Presets
+        if stream_name not in Presets().stream_presets.keys():
+            self.app.create_preset(stream_name, preset_type, num_channels=num_channels, nominal_sample_rate=srate, **kwargs)
         else:
             raise ValueError(f"test_utils.create_add_start_stream(): Stream name {stream_name} is already in keys for app.presets")
 
@@ -238,7 +250,7 @@ class ContextBot:
 
         self.qtbot.mouseClick(self.app.addStreamWidget.add_btn, QtCore.Qt.MouseButton.LeftButton)  # click the add widget combo box
 
-    def create_zmq_stream(self, stream_name: str, num_channels: int, srate:int, port_range=(5000, 5100), data_type=DataType.uint8):
+    def create_zmq_stream(self, stream_name: str, port_range=(5000, 5100), **kwargs):
         from physiolabxr.sub_process.pyzmq_utils import can_connect_to_port
         using_port = None
         for port in range(*port_range):
@@ -249,7 +261,7 @@ class ContextBot:
             raise ValueError(f"Could not find a port in range {port_range}. Consider use a different range.")
         if stream_name in self.send_data_processes.keys():
             raise ValueError(f"Stream name {stream_name} is in keys for send_data_processes")
[…]
