### 2023-10-28 af57445
update test utils
