### 2023-10-29 #5854 @ApocalyVec
> Hi @ApocalyVec, thanks for this nice work. The paper has a few references to 'real-time'. I consider this a 'performance' claim as per the checkbox above. Could you please elaborate in the paper on what is meant by real-time (milliseconds, microseconds as in an RTOS, etc.)? Further what has been done to verify the real-time performance?
>
> Second, the paper mentions closed-loop control: lines 26-27 "often involving closed-loop interaction systems". Has the software been set up to mediate a closed-loop configuration? If so what sensors and actuators were used and what latency did the software support?
>
> I post these comments here, rather than as an issue on the source repository, since each is very specific to the JOSS paper.

Hi @Lucask07, thank you for your review and constructive feedback, and I apologize it took a while to get back to you. To answer your questions,

PhysioLabXR is a soft real-time system. On the main thread, the best case is when all data is pulled from the network/device buffer before new data becomes available. Otherwise, data will accumulate in their buffers, leading to deteriorating performance. Within PhysioLabXR, we do the following to get close to a smooth run time:

- **Pull data**: the operation collects all available data at once. The acquired data are processed by the following subroutines, each run on a different timer. Not all of them may be running depending on the user’s task. In the system, the pull-data threads have the highest priorities.
- **Visualization**: displaying the data in the GUI. Default at 30 Hz. Viz frame-per-second is influenced by the amount of data that needs to be updated between frames (e.g., the number of channels, image resolution, etc.)
- **Real-time filtering**: the real-time filtering happens right after the data has been pulled. Filter parameters such as order and buffer size can affect the overhead it adds.
- **Recording**: evicting the recording buffer to disk. Default at 1 Hz.
- **Forward input data to user scripts**: run frequency depends on user settings.

The actual runtime performance of PhysioLabXR varies depending on the specs of the host computer, other running tasks competing for CPU time, and the amount of data that the user is streaming.

We added to the docs: benchmarks across data throughputs commonly found in physiological experiments as a de facto verification of the performance.

[https://physiolabxrdocs.readthedocs.io/en/latest/Benchmarks.html](https://physiolabxrdocs.readthedocs.io/en/latest/Benchmarks.html)

The metrics in the benchmarks measure the running frequencies of 1) visualization and 2) pull data. This is because these two subroutines sit at the two sides of the spectrum in terms of priority. The plotting is the lowest priority thread in PhysioLabXR because missing the deadline has the least impact on the system’s integrity (though it affects user experience).

We hope including these benchmarks can give users a hint at what performance level they should expect given their experiment setup.
