### 2023-10-29 #5854 @ApocalyVec
> @ApocalyVec
>
> Could you please walk us through the checkbox for "**Human and animal research**: If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item."
>
> I see an example dataset such as [Eye fixation data](https://physiolabxrdocs.readthedocs.io/en/latest/FixationDetection.html) and am not sure if this data is relevant to the policy and if it is relevant if the research complies with the JOSS policy.
>
> Fig. 2 of the manuscript also seems to have human subject data. Please help us understand need for compliance here as well.

@lucask07 @mstimberg Thanks for raising the concern about data policy from human subjects. Let me clarify what data is in the paper. We obtained consent before collecting the data, and it is deidentified in the form that is presented (the data shown in the paper was, in fact, contributed by one of the co-authors). All the procedures we carried out while developing the software are in accordance with the Declaration of Helsinki.
