### 2023-10-29 2f2a6dc
fix grammar issues and fix sentences per @lucask07's suggestions
