+++ paper/paper.bib
@@ -44,6 +44,14 @@
   note = {Accessed: May 10, 2023}
 }
 
+@misc{neuropixels,
+  title = {Neuropixels},
+  author = {{Interuniversity Microelectronics Centre}},
+  year = {2023},
+  url = {https://www.neuropixels.org/},
+  note = {Accessed: Oct 28, 2023}
+}
+
 @misc{matlab,
   title = {{MATLAB}},
   note = {Version R2021b},
