### 2023-10-29 #5854 @ApocalyVec
> [lucask07](/lucask07)

Hey @lucask07 really appreciate catching these mistakes! Just pushed the fix, asking the bot to compile a new proof now!
