### 2023-10-29 #5854 @ApocalyVec
> @ApocalyVec a few paper grammar notes:
>
> 1. Line 87 extra space
> 2. Line 142 stray period after maximizing (?)
> 3. Line 151 to 154: unclear / awkward sentence

Hey @lucask07 really appreciate catching these mistakes! Just pushed the fix, asking the bot to compile a new proof now!
