### 2023-10-29 #5854 @ApocalyVec
> @ApocalyVec Thank you very much for your excellent work. I have some comments and questions as follows:
>
> Document: 1- The section "Python API, Write your data source script" doesn't have any content. Also, it seems that there is no solution or explanation for visualizing the data recorded in this way.
>
> 2- I can't see any API reference in the documents (No document in the code). API documentation are very helpful in tutorials/examples.
>
> 3- This tutorial, "Building ERP Classifier based on EEG and Pupil Size" is not clear.
>
> paper:
>
> 1- It's better to keep the figures' captions short and explain them in the main body. 2- I suggest adding a conclusion section to the paper, including software limitations, future work, and a summary. 3- In several parts of the paper, the authors claimed high performance and rapid prototyping for developing experiments. I suggest comparing its performance with similar software or libraries like [PsychoPy](https://www.psychopy.org/) and [Octopus Sensing](https://octopus-sensing.nastaran-saffar.me/) 4- I can see this software has many features similar to [Octopus Sensing](https://octopus-sensing.nastaran-saffar.me/) (related paper in JOSS: https://joss.theoj.org/papers/10.21105/joss.04045). It may be good to discuss why you chose not to integrate with or extend it.
>
> Software: 1-Usually, in EEG-based studies, we need to send a marker or trigger to mark an event on data between start and stop. It seems that PhysioLabXR doesn't support this. or maybe it records markers in a separate file from the sensor data. (In this case, we need a script to match it with data and know when exactly the event happened, for example, visualization or processing). Could you please explain if it is possible to add it to PhysioLabXR, or how do you handle this currently? I also suggest adding it as an issue to the repository to resolve in the future.
>
> 2- It seems that LSL has been used for data recording from any sensor. If so, the system performance depends on the network and the other software for reading data from a sensor. Could you please make it more clear or discuss it?
>
> Installation 1- I installed it with pip. It was done successfully. However, it had some errors in pyaudio installation (Ubuntu OS).

Thank you, @nastaran62, for your thorough review, really appreciate it! We are working through the items you brought up and will get back to you by the end of the coming week.
