+++ physiolabxr/examples/coe_video/send_fmri_data_example/FMRIZMQExampleOutlet.py
@@ -7,7 +7,6 @@ import numpy as np
 import zmq
 from pylsl import local_clock
 import pickle
-from physiolabxr.examples.fmri_experiment_example.mri_utils import load_nii_gz_file
 from scipy.ndimage import zoom
 
 def main():
@@ -45,25 +44,33 @@ def main():
     start_time = time.time()
     sent_samples = 0
 
-    image = fmri_data[:, :, :, 10]
+    image_list = []
+
+    for i in range(12):
+        image = fmri_data[:, :, :, i]
+        # move the image down
+
+
+        new_matrix = image[:, :, 17:]
+
+        # Pad 5 zero layers to the top
+        padding = np.zeros((256, 256, 17))
+        image = np.concatenate((new_matrix, padding), axis=-1)
+        # image = (image - np.min(image)) / (np.max(image) - np.min(image))
+        # zoom image to
+        scale_factor = (0.9375, 0.9375, 1.5)
+
+        # Rescale the matrix using zoom
+        image = zoom(image, scale_factor)
+        image_list.append(image)
+    image_list = np.array(image_list)
+    image_list = (image_list - np.min(image_list)) / (np.max(image_list) - np.min(image_list))
+
     del fmri_data
-    # do the image down
-    new_matrix = image[:, :, 17:]
-
-    # Pad 5 zero layers to the top
-    padding = np.zeros((256, 256, 17))
-    image = np.concatenate((new_matrix, padding), axis=-1)
-    del new_matrix
-    image = (image - np.min(image)) / (np.max(image) - np.min(image))
-    # zoom image to
-    scale_factor = (0.9375, 0.9375, 1.5)
-
-    # Rescale the matrix using zoom
-    image = zoom(image, scale_factor)
 
     #
     # image_array = image.flatten()
-
+    sent_sample = 0
     while True:
         elapsed_time = time.time() - start_time
         required_samples = int(srate * elapsed_time) - sent_samples
@@ -83,7 +90,11 @@ def main():
                 # padding = np.zeros((256, 256, 5))
                 # image = np.concatenate((padding, new_matrix), axis=-1)
 
-                image_array = image.flatten()
+                # image_index = random.randint(0, len(image_list)-1)
+                # print(image_index)
+                image_index = sent_sample % len(image_list)
+                sent_sample += 1
+                image_array = image_list[image_index].flatten()
 
                 mysample = image_array
                 socket.send_multipart([bytes(topic, "utf-8"), np.array(local_clock()), mysample])
