+++ tests/test_utils.py
@@ -129,7 +129,7 @@ class ContextBot:
     def cleanup(self):
         pass
 
-    def create_add_start_stream(self, stream_name: str, num_channels: int, srate:int, after_process_wait_time=1):
+    def create_add_start_stream(self, stream_name: str, num_channels: int, srate:int, start_stream=True):
         """
         start a stream as a separate process, add it to the app's streams, and start it once it becomes
         available
@@ -151,8 +151,9 @@ class ContextBot:
         self.qtbot.keyClicks(self.app.addStreamWidget.stream_name_combo_box, stream_name)
         self.qtbot.mouseClick(self.app.addStreamWidget.add_btn, QtCore.Qt.MouseButton.LeftButton)  # click the add widget combo box
 
-        self.qtbot.waitUntil(lambda: stream_is_available(app=self.app, test_stream_name=stream_name), timeout=self.stream_availability_timeout)  # wait until the LSL stream becomes available
-        self.qtbot.mouseClick(self.app.stream_widgets[stream_name].StartStopStreamBtn, QtCore.Qt.MouseButton.LeftButton)
+        if start_stream:
+            self.qtbot.waitUntil(lambda: stream_is_available(app=self.app, test_stream_name=stream_name), timeout=self.stream_availability_timeout)  # wait until the LSL stream becomes available
+            self.qtbot.mouseClick(self.app.stream_widgets[stream_name].StartStopStreamBtn, QtCore.Qt.MouseButton.LeftButton)
 
     def add_and_start_stream(self, stream_name: str, num_channels:int, interface_type=PresetType.LSL, port=None, dtype: DataType=None, *args, **kwargs):
         self.add_stream(stream_name, interface_type, port, dtype, *args, **kwargs)
@@ -349,7 +350,13 @@ def run_visualization_benchmark(app_main_window, test_context, test_stream_names
         print(f"Testing #channels {num_channels} and srate {sampling_rate} with random stream name(s) {stream_names}...", end='')
         start_time = time.perf_counter()
         for s_name in stream_names:
-            test_context.create_add_start_stream(s_name, num_channels, sampling_rate)
+            test_context.create_add_start_stream(s_name, num_channels, sampling_rate, start_stream=False)
+
+        # wait some time for streams to become available
+        test_context.qtbot.wait(int(4 * 1e3))
+        for s_name in stream_names:
+            test_context.start_a_stream(s_name, num_channels, thread_timer_second=4)
+
         if is_reocrding:
             app_main_window.settings_widget.set_recording_file_location(os.getcwd())  # set recording file location (not through the system's file dialog)
             test_context.qtbot.mouseClick(app_main_window.recording_tab.StartStopRecordingBtn, QtCore.Qt.MouseButton.LeftButton)  # start the recording
@@ -409,8 +416,12 @@ def plot_viz_benchmark_results(results, test_axes, metrics, notes=''):
     num_channels_to_test = test_axes["number of channels"]
     num_streams_to_test = test_axes["number of streams"]
 
+    metric_values = {metric_name: [value_dict[metric_name] for test_combo, value_dict in result_dict.items() ] for metric_name, result_dict in results.items()}
+    metric_vmin_max = {metric_name: (np.min(values), np.max(values)) for metric_name, values in metric_values.items()}
+
     for n_streams in num_streams_to_test:
-        visualize_metrics_across_num_chan_sampling_rate(results, metrics, n_streams, sampling_rates_to_test, num_channels_to_test, notes=notes)
+
+        visualize_metrics_across_num_chan_sampling_rate(results, metrics, n_streams, sampling_rates_to_test, num_channels_to_test, notes=notes, metric_vmin_max=metric_vmin_max)
 
[…]
