### 2023-10-29 bfbff34
add notification pane
