### 2023-10-30 f527c2c
update demo scripts and simulated target
