+++ physiolabxr/examples/coe_video/RenaScript/Model.py
@@ -1,0 +1,12 @@
+import numpy as np
+
+from physiolabxr.scripting.RenaScript import RenaScript
+
+
+class Model():
+
+    def __init__(self):
+        pass
+
+    def pred(self, pupil, fmri):
+        return np.random.rand(1, 1)
