+++ physiolabxr/examples/physio_helpers.py
@@ -1,0 +1,5 @@
+def detect_fixations(*args, **kwargs):
+    return []
+
+def load_model(*args, **kwargs):
+    return []
