### 2023-10-30 6e36b04
clear buffer up to train end
