### 2023-10-30 95bc56e
clear buffer up to train end
