### 2023-11-01 70a28fd
add params in motor imagery
