### 2023-11-06 PhysioLabXR/PhysioLabXR#120 @lucask07
type object not subscriptable error with Python3.8 
I received the error below when installing in a conda environment with Python3.8. This error did not occur with Python3.10.

```
  File "/Users/koer2434/opt/anaconda3/envs/phys2/lib/python3.8/site-packages/physiolabxr/presets/Presets.py", line 14, in <module>
    from physiolabxr.presets.GroupEntry import GroupEntry
  File "/Users/koer2434/opt/anaconda3/envs/phys2/lib/python3.8/site-packages/physiolabxr/presets/GroupEntry.py", line 12, in <module>
    from physiolabxr.utils.dsp_utils.dsp_modules import *
  File "/Users/koer2434/opt/anaconda3/envs/phys2/lib/python3.8/site-packages/physiolabxr/utils/dsp_utils/dsp_modules.py", line 344, in <module>
    def run_data_processors(data, data_processor_pipeline: list[DataProcessor]):
TypeError: 'type' object is not subscriptable
```
