### 2023-11-06 #5854 @lucask07
@ApocalyVec
Installation works in a conda environment with Python3.10. I think the package should state that Python must be >3.8. See issue https://github.com/PhysioLabXR/PhysioLabXR/issues/120
