### 2023-11-06 736e771
Add illumiReadSwypeScript.py finish state transition TODO: work on the decoding process
