### 2023-11-06 #5854 @mstimberg
> @lucask07 @mstimberg Thanks for raising the concern about data policy from human subjects. Let me clarify what data is in the paper. We obtained consent before collecting the data, and it is deidentified in the form that is presented (the data shown in the paper was, in fact, contributed by one of the co-authors). All the procedures we carried out while developing the software are in accordance with the Declaration of Helsinki.

Thanks for the clarifications. Did these recordings (and the EEG recordings here: https://physiolabxrdocs.readthedocs.io/en/latest/tutorials/BuildMultiModalERPClassifier.html) require ethics committee approval? If yes, could you please state this in the usual way in the manuscript (reference number, etc.)? If not, e.g. if all the data are recordings from the co-authors (not 100% sure whether this changes things), please do state this as well. Of course, if these data are already published elsewhere, you can also simply cite these references. Thanks :pray: !
