### 2023-11-06 PhysioLabXR/PhysioLabXR#120 @ApocalyVec
Hey @lucask07, thank you for pointing this out. Yes indeed subscription in type is only added in Python 3.9+ as per [the version notes here](https://docs.python.org/3.9/whatsnew/3.9.html#type-hinting-generics-in-standard-collections), and [the discussion](https://stackoverflow.com/questions/75202610/typeerror-type-object-is-not-subscriptable-python) here.

We updated the note on Python versions in the docs: [install with pip](https://physiolabxrdocs.readthedocs.io/en/latest/#install-with-pip), and [run from source](https://physiolabxrdocs.readthedocs.io/en/latest/Contribute.html#run-from-source).
