### 2023-11-06 #5854 @ApocalyVec
> @ApocalyVec Installation works in a conda environment with Python3.10. I think the package should state that Python must be >3.8. See issue [PhysioLabXR/PhysioLabXR#120](https://github.com/PhysioLabXR/PhysioLabXR/issues/120)

Hey @lucask07 , thanks for the note on Python versions. Just reply to [the issue you raise in the repo](https://github.com/PhysioLabXR/PhysioLabXR/issues/120).
