### 2023-11-06 33c70e8
dummy test only open main window
