+++ .github/workflows/unit-tests.yml
@@ -9,11 +9,12 @@ jobs:
     strategy:
       matrix:
         os : [ubuntu-latest]
-        python-version: [ "3.9" ]
+        python-version: [ "3.10" ]
     env:
       DISPLAY: ':99.0'
     steps:
-      - uses: actions/checkout@v3
+      - name: checkout repo
+        uses: actions/checkout@v3
       - uses: tlambert03/setup-qt-libs@v1
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
