### 2023-11-06 e9274f2
dummy test only open main window
