+++ tests/Dtest.py
@@ -30,16 +30,12 @@ def app_main_window(qtbot):
 
 @pytest.fixture
 def context_bot(app_main_window, qtbot):
+    print("here")
     test_context = ContextBot(app=app_main_window, qtbot=qtbot)
     yield test_context
     test_context.clean_up()
 
 
-@pytest.fixture
-def context_bot(app_main_window, qtbot):
-    test_context = ContextBot(app=app_main_window, qtbot=qtbot)
-    yield test_context
-    test_context.clean_up()
 
 
 def test_add_inactive_unknown_stream_in_added_stream_widgets(app_main_window, qtbot) -> None:
