+++ .github/workflows/unit-tests.yml
@@ -23,11 +23,17 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          sudo apt install libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 x11-utils libegl1
-          /sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
+          sudo apt install -y xvfb x11-utils libxkbcommon-x11-0
           sudo apt install qt6-base-dev
+          pip install pyside2 pytest coverage pytest-qt
           pip install wheel
           pip install -r requirements.dev.txt
-      - name: Test with pytest
+      - name: Test
+        env:
+          QT_DEBUG_PLUGINS: 1
         run: |
-          ./tests/test.sh
+          ulimit -c unlimited
+          sudo catchsegv xvfb-run --auto-servernum `which coverage` run -m pytest ./tests/test.sh
+          
+
+          
