+++ .github/workflows/unit-tests.yml
@@ -32,7 +32,6 @@ jobs:
         env:
           QT_DEBUG_PLUGINS: 1
         run: |
-          ulimit -c unlimited
           sudo catchsegv xvfb-run --auto-servernum `which coverage` run -m pytest ./tests/test.sh
           
 
