+++ tests/Dtest.py
@@ -11,7 +11,7 @@ from PyQt6 import QtCore
 from PyQt6.QtCore import Qt
 from PyQt6.QtWidgets import QWidget
 from physiolabxr.configs.configs import AppConfigs
-AppConfigs(_reset=True)  # create the singleton app configs object
+#AppConfigs(_reset=True)  # create the singleton app configs object
 
 from tests.TestStream import LSLTestStream
 from tests.test_utils import app_fixture, ContextBot, get_random_test_stream_names
