+++ .github/workflows/unit-tests.yml
@@ -28,8 +28,13 @@ jobs:
           sudo apt install libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 x11-utils libegl1
           /sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
           sudo apt install qt6-base-dev
+          sudo apt install -y xvfb x11-utils libxkbcommon-x11-0
           pip install wheel
           pip install -r requirements.dev.txt
+          pip install pyside2 pytest coverage pytest-qt
       - name: Test with pytest
+        env:
+          QT_DEBUG_PLUGINS: 1
         run: |
-          ./tests/test.sh 
+          sudo xvfb-run  ./tests/test.sh
+           
