+++ .github/workflows/unit-tests.yml
@@ -33,8 +33,8 @@ jobs:
           pip install -r requirements.dev.txt
           pip install pyside2 pytest coverage pytest-qt
       - name: Test with pytest
-        env:
-          QT_DEBUG_PLUGINS: 1
-        run: |
-          sudo xvfb-run  ./tests/test.sh
+        uses: coactions/setup-xvfb@v1
+        with:
+          run: |
+            ./tests/test.sh
            
