+++ .github/workflows/unit-tests.yml
@@ -16,8 +16,6 @@ jobs:
       - name: checkout repo
         uses: actions/checkout@v3
       - uses: tlambert03/setup-qt-libs@v1
-      - name: Install Qt
-        uses: jurplel/install-qt-action@v3
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
@@ -31,7 +29,7 @@ jobs:
           sudo apt install -y xvfb x11-utils libxkbcommon-x11-0
           pip install wheel
           pip install -r requirements.dev.txt
-          pip install pyside2 pytest coverage pytest-qt
+
       - name: Test with pytest
         uses: coactions/setup-xvfb@v1
         with:
