+++ .github/workflows/unit-tests.yml
@@ -15,7 +15,8 @@ jobs:
     steps:
       - name: checkout repo
         uses: actions/checkout@v3
-      - uses: tlambert03/setup-qt-libs@v1
+      - name: Install Qt
+        uses: jurplel/install-qt-action@v3
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
@@ -23,9 +24,6 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          sudo apt install libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 x11-utils libegl1
-          /sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
-          sudo apt install qt6-base-dev
           sudo apt install -y xvfb x11-utils libxkbcommon-x11-0
           pip install wheel
           pip install -r requirements.dev.txt
@@ -34,6 +32,5 @@ jobs:
         uses: coactions/setup-xvfb@v1
         with:
           run: |
-            echo "Running tests..."
             ./tests/test.sh
            
