+++ .github/workflows/unit-tests.yml
@@ -25,6 +25,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           sudo apt install -y xvfb x11-utils libxkbcommon-x11-0
+          sudo apt-get install '^libxcb.*-dev' libx11-xcb-dev libglu1-mesa-dev libxrender-dev libxi-dev libxkbcommon-dev libxkbcommon-x11-dev
           pip install wheel
           pip install -r requirements.dev.txt
           pip uninstall opencv-python -y
