+++ .github/workflows/unit-tests.yml
@@ -36,5 +36,6 @@ jobs:
           env:
             QT_DEBUG_PLUGINS:1
           run: |
+            echo "Starting test"
             ./tests/test.sh
            
