+++ .github/workflows/unit-tests.yml
@@ -21,6 +21,12 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
+      - name: Setup xvfb (Linux)
+        if: runner.os == 'Linux'
+        run: |
+          sudo apt-get install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xinput0 libxcb-xfixes0
+          # start xvfb on (default) display :0
+          sudo /usr/bin/Xvfb $DISPLAY -screen 0 1280x1024x24 &
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
@@ -30,11 +36,12 @@ jobs:
           pip install -r requirements.dev.txt
           pip uninstall opencv-python -y
           pip install opencv-python-headless
+          pip install pyside2 pytest coverage pytest-qt
+          python -m pip install codecov coverage flake8 pytest
+      - name: Start application and print version
+        run: |
+          QT_DEBUG_PLUGINS=1 python -m shapeout2 --version
       - name: Test with pytest
-        uses: coactions/setup-xvfb@v1
-        with:
-          env:
-            QT_DEBUG_PLUGINS:1
-          run: |
-            npm ./tests/test.sh
+        run: |
+         ./tests/test.sh
            
