### 2023-11-07 #5854 @mstimberg
> Hey @mstimberg, thanks for raising the issue. To answer your question, yes, these data (EEG, eyetracking, etc.) are all from co-authors and unpublished except what we shared in the docs.

Great, thanks for confirming. Could you please add a little note stating this to the paper (probably best it in the "Licensing and Availability" section)? :pray:
