+++ .github/workflows/unit-tests.yml
@@ -21,12 +21,10 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Setup xvfb (Linux)
-        if: runner.os == 'Linux'
+      - name: Start Xvfb
         run: |
           sudo apt-get install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xinput0 libxcb-xfixes0
-          # start xvfb on (default) display :0
-          sudo /usr/bin/Xvfb $DISPLAY -screen 0 1280x1024x24 &
+          Xvfb :99 -ac -screen 0 1024x768x24 & export DISPLAY=:99
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
@@ -37,11 +35,11 @@ jobs:
           pip uninstall opencv-python -y
           pip install opencv-python-headless
           pip install pyside2 pytest coverage pytest-qt
-          python -m pip install codecov coverage flake8 pytest
+          python -m pip install coverage pytest
       - name: Start application and print version
         run: |
           QT_DEBUG_PLUGINS=1
       - name: Test with pytest
         run: |
-         ./tests/test.sh
+         xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" coverage run ./tests/test.sh
            
