+++ .github/workflows/unit-tests.yml
@@ -16,7 +16,7 @@ jobs:
       - name: checkout repo
         uses: actions/checkout@v3
       - name: Install Qt
-        uses: tlambert03/setup-qt-libs
+        uses: tlambert03/setup-qt-libs@v1
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
