+++ .github/workflows/unit-tests.yml
@@ -36,8 +36,10 @@ jobs:
           pip install opencv-python-headless
           pip install pyside2 pytest coverage pytest-qt
           python -m pip install coverage pytest
+          pip install pytest-xvfb
 
       - name: Test with pytest
         run: |
          xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" ./tests/test.sh
-           
+
+
