+++ .github/workflows/unit-tests.yml
@@ -24,7 +24,6 @@ jobs:
       - name: Start Xvfb
         run: |
           sudo apt-get install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xinput0 libxcb-xfixes0
-          Xvfb :99 -ac -screen 0 1024x768x24 & export DISPLAY=:99
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
@@ -41,10 +40,9 @@ jobs:
 
       - name: Check job values
         run: |
-            apt list PyQt6 --installed
-            qtdiag
+            qtchooser -list-versions
             QT_SELECT=qt6
-            qtdiag
+
       - name: Test with pytest
         run: |
          xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" ./tests/test.sh
