### 2023-11-09 44db811
qmake --version
