+++ .github/workflows/unit-tests.yml
@@ -40,7 +40,7 @@ jobs:
 
       - name: Check job values
         run: |
-            qtchooser -list-versions
+            qmake --version
             QT_SELECT=qt6
 
       - name: Test with pytest
