### 2023-11-09 7f3eb58
qmake --version
