### 2023-11-09 af87d0e
qmake --version
