### 2023-11-09 b6d0d33
qmake --version
