### 2023-11-09 PhysioLabXR/PhysioLabXR#120 @ApocalyVec
Closing this issue setting the minimal requirements for PhysioLabXR to be Python 3.9
