+++ .github/workflows/unit-tests.yml
@@ -43,6 +43,6 @@ jobs:
 
       - name: Test with pytest
         run: |
-         xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" ./tests/test.sh
+         sudo xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" ./tests/test.sh
 
 
