+++ .github/workflows/unit-tests.yml
@@ -29,13 +29,14 @@ jobs:
           python -m pip install --upgrade pip
           sudo apt install -y xvfb x11-utils libxkbcommon-x11-0
           sudo apt-get install '^libxcb.*-dev' libx11-xcb-dev libglu1-mesa-dev libxrender-dev libxi-dev libxkbcommon-dev libxkbcommon-x11-dev
-          sudo pip install wheel
+          pip install wheel
           pip install -r requirements.dev.txt
           pip uninstall opencv-python -y
           pip install opencv-python-headless
           sudo pip install PyQt6
           sudo pip install pytest
           sudo pip install pytest-qt
+          pip install pyscreeze
 
       - name: Check job values
         run: |
