+++ pytest.ini
@@ -1,0 +1,6 @@
+# pytest.ini
+[pytest]
+testpaths =
+    tests
+python_files = test_*.py
+addopts = -rf --import-mode=importlib
