+++ .github/workflows/unit-tests.yml
@@ -33,7 +33,7 @@ jobs:
           pip install -r requirements.dev.txt
           pip uninstall opencv-python -y
           pip install opencv-python-headless
-          pip install PyQt6
+          sudo pip install PyQt6
           sudo pip install pytest
           sudo pip install pytest-qt
           pip install PyScreeze
