+++ .github/workflows/unit-tests.yml
@@ -29,14 +29,14 @@ jobs:
           python -m pip install --upgrade pip
           sudo apt install -y xvfb x11-utils libxkbcommon-x11-0
           sudo apt-get install '^libxcb.*-dev' libx11-xcb-dev libglu1-mesa-dev libxrender-dev libxi-dev libxkbcommon-dev libxkbcommon-x11-dev
-          pip install wheel
-          sudo pip install -r requirements.dev.txt
-          pip uninstall opencv-python -y
-          pip install opencv-python-headless
-          sudo pip install PyQt6
-          sudo pip install pytest
-          sudo pip install pytest-qt
-          pip install PyScreeze
+          python -m pip install wheel
+          python -m pip install -r requirements.dev.txt
+          python -m pip uninstall opencv-python -y
+          python -m pip install opencv-python-headless
+          python -m pip install PyQt6
+          python -m pip install pytest
+          python -m pip install pytest-qt
+          python -m pip install PyScreeze
 
       - name: Check job values
         run: |
