+++ .github/workflows/unit-tests.yml
@@ -36,7 +36,6 @@ jobs:
           python -m pip install PyQt6
           python -m pip install pytest
           python -m pip install pytest-qt
-          python -m pip install PyScreeze
 
       - name: Check job values
         run: |
@@ -44,6 +43,6 @@ jobs:
 
       - name: Test with pytest
         run: |
-         xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" ./tests/test.sh
+         xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" pytest RenaVisualizationTest.py -s 
 
 
