+++ .github/workflows/unit-tests.yml
@@ -33,9 +33,9 @@ jobs:
           python -m pip install -r requirements.dev.txt
           python -m pip uninstall opencv-python -y
           python -m pip install opencv-python-headless
-          python -m pip install PyQt6
-          python -m pip install pytest
-          python -m pip install pytest-qt
+          sudo python -m pip install PyQt6
+          sudo python -m pip install pytest
+          sudo python -m pip install pytest-qt
 
       - name: Check job values
         run: |
@@ -43,6 +43,6 @@ jobs:
 
       - name: Test with pytest
         run: |
-         xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" pytest RenaVisualizationTest.py -s 
+         sudo xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" pytest RenaVisualizationTest.py -s 
 
 
