+++ .github/workflows/unit-tests.yml
@@ -43,6 +43,6 @@ jobs:
 
       - name: Test with pytest
         run: |
-         sudo xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" pytest ./tests/RenaVisualizationTest.py -s 
+         sudo xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" pytest ./tests/SetupTest.py -s 
 
 
