+++ tox.ini
@@ -1,0 +1,19 @@
+[tox]
+minversion = 3.8.0
+envlist = py36, py37, py38, py39, flake8, mypy
+isolated_build = true
+
+[gh-actions]
+python =
+    3.6: py36
+    3.7: py37
+    3.8: py38
+    3.9: py39
+
+[testenv]
+setenv =
+    PYTHONPATH = {toxinidir}
+deps =
+    -r{toxinidir}/requirements_dev.txt
+commands =
+    pytest tests/SetupTest.py -s
