+++ tox.ini
@@ -14,6 +14,6 @@ python =
 setenv =
     PYTHONPATH = {toxinidir}
 deps =
-    -r{toxinidir}/requirements_dev.txt
+    -r{toxinidir}/requirements.dev.txt
 commands =
     pytest tests/SetupTest.py -s
