+++ tox.ini
@@ -1,6 +1,5 @@
 [tox]
-minversion = 3.8.0
-envlist = py36, py37, py38, py39, flake8, mypy
+envlist = py{37,38,39,310}-{pyqt5,pyside2,pyside6,pyqt6}
 isolated_build = true
 
 [gh-actions]
@@ -13,7 +12,14 @@ python =
 [testenv]
 setenv =
     PYTHONPATH = {toxinidir}
+    PYTEST_QT_API=pyqt6
+    QT_QPA_PLATFORM=offscreen
 deps =
     -r{toxinidir}/requirements.dev.txt
+    pyqt6: pyqt6
+passenv=
+    DISPLAY
+    XAUTHORITY
+    COLUMNS
 commands =
     pytest tests/SetupTest.py -s
