+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py{37,38,39,310}-{pyqt5,pyside2,pyside6,pyqt6}
+envlist = py{310}-{pyqt6}
 isolated_build = true
 
 [gh-actions]
