+++ .github/workflows/unit-tests.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       matrix:
         os : [ubuntu-latest]
-        python-version: [ "3.10" ]
+        python-version: [ "3.9" ]
     env:
       DISPLAY: ':99.0'
     steps:
