+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py{310}-{pyqt6}
+envlist = py{39}-{pyqt6}
 isolated_build = true
 
 [gh-actions]
