### 2023-11-10 40147a1
deleted requirements.dev.txt --> opencv-python
