+++ tox.ini
@@ -1,12 +1,8 @@
 [tox]
 envlist = py{39}-{pyqt6}
-isolated_build = true
 
 [gh-actions]
 python =
-    3.6: py36
-    3.7: py37
-    3.8: py38
     3.9: py39
 
 [testenv]
