+++ .github/workflows/unit-tests.yml
@@ -33,9 +33,9 @@ jobs:
           python -m pip install -r requirements.dev.txt
           python -m pip uninstall opencv-python -y
           python -m pip install opencv-python-headless
-          sudo python -m pip install PyQt6
-          sudo python -m pip install pytest
-          sudo python -m pip install pytest-qt
+          python -m pip install PyQt6
+          python -m pip install pytest
+          python -m pip install pytest-qt
           pip install tox
 
       - name: Check job values
