+++ .github/workflows/unit-tests.yml
@@ -44,4 +44,5 @@ jobs:
 
       - name: Test with pytest
         run: |
-         tox
+         sudo xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" ./tests/test.sh -s 
+         
