+++ tests/Dtest.py
@@ -54,7 +54,7 @@ def test_add_inactive_unknown_stream_in_added_stream_widgets(app_main_window, qt
     test_stream_name = 'TestStreamName'
 
     app_main_window.ui.tabWidget.setCurrentWidget(app_main_window.ui.tabWidget.findChild(QWidget, 'visualization_tab'))  # switch to the visualization widget
-    # qtbot.mouseClick(app_main_window.addStreamWidget.stream_name_combo_box, QtCore.Qt.MouseButton.LeftButton)  # click the add widget combo box
+    qtbot.mouseClick(app_main_window.addStreamWidget.stream_name_combo_box, QtCore.Qt.MouseButton.LeftButton)  # click the add widget combo box
     # qtbot.keyPress(app_main_window.addStreamWidget.stream_name_combo_box, Qt.Key.Key_A, modifier=Qt.KeyboardModifier.ControlModifier)
     # qtbot.keyClicks(app_main_window.addStreamWidget.stream_name_combo_box, test_stream_name)
     # assert app_main_window.addStreamWidget.add_btn.isEnabled()
