+++ pytest.ini
@@ -3,4 +3,3 @@
 testpaths =
     tests
 python_files = test_*.py
-addopts = -rf --import-mode=importlib
