+++ .github/workflows/unit-tests.yml
@@ -29,8 +29,8 @@ jobs:
           python -m pip install --upgrade pip
           sudo apt install -y xvfb x11-utils libxkbcommon-x11-0
           sudo apt-get install '^libxcb.*-dev' libx11-xcb-dev libglu1-mesa-dev libxrender-dev libxi-dev libxkbcommon-dev libxkbcommon-x11-dev
-          python -m pip install wheel
-          python -m pip install -r requirements.dev.txt
+          sudo python -m pip install wheel
+          sudo python -m pip install -r requirements.dev.txt
           python -m pip uninstall opencv-python -y
           python -m pip install opencv-python-headless
           sudo pip install PyQt6
