+++ .github/workflows/unit-tests.yml
@@ -24,6 +24,7 @@ jobs:
       - name: Start Xvfb
         run: |
           sudo apt-get install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xinput0 libxcb-xfixes0
+          xvfb-run --auto-servernum --server-num=1 --server-args="-screen 1 1920x1200x24 -ac +extension GLX" echo "Xvfb started"
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
@@ -40,7 +41,7 @@ jobs:
 
       - name: Check job values
         run: |
-            QT_SELECT=qt6
+           export QT_SELECT=qt6
 
       - name: Test with pytest
         run: |
