+++ pytest.ini
@@ -3,3 +3,7 @@
 testpaths =
     tests
 python_files = test_*.py
+timeout = 100
+log_cli=true
+log_level=DEBUG
+console_output_style = progress-even-when-capture-no
