+++ physiolabxr/scripting/illumiRead/illumiReadSwype/illumiReadSwypeScript.py
@@ -1,0 +1,188 @@
+import time
+from collections import deque
+import cv2
+import numpy as np
+import time
+import os
+import pickle
+import sys
+import matplotlib.pyplot as plt
+from physiolabxr.scripting.RenaScript import RenaScript
+from pylsl import StreamInfo, StreamOutlet, cf_float32
+import torch
+from physiolabxr.scripting.illumiRead.illumiReadSwype import illumiReadSwypeConfig
+from physiolabxr.scripting.illumiRead.illumiReadSwype.illumiReadSwypeConfig import EventMarkerLSLStreamInfo, \
+    GazeDataLSLStreamInfo
+
+
+class AOIAugmentationScript(RenaScript):
+    def __init__(self, *args, **kwargs):
+        """
+        Please do not edit this function
+        """
+        super().__init__(*args, **kwargs)
+
+        self.currentExperimentState: illumiReadSwypeConfig.ExperimentState = \
+            illumiReadSwypeConfig.ExperimentState.InitState
+
+        self.currentBlock: illumiReadSwypeConfig.ExperimentBlock = \
+            illumiReadSwypeConfig.ExperimentBlock.InitBlock
+
+        self.device = torch.device('cpu')
+
+        self.current_image_name = None
+
+        self.process_gaze_data_time_buffer = deque(maxlen=1000)
+
+    def init(self):
+        pass
+
+    # loop is called <Run Frequency> times per second
+    def loop(self):
+
+        if (EventMarkerLSLStreamInfo.StreamName not in self.inputs.keys()) or (
+                GazeDataLSLStreamInfo.StreamName not in self.inputs.keys()):  # or GazeDataLSLOutlet.StreamName not in self.inputs.keys():
+            return
+        # print("process event marker call start")
+        self.process_event_markers()
+
+    def cleanup(self):
+        print('Cleanup function is called')
+
+    def process_event_markers(self):
+        event_markers = self.inputs[EventMarkerLSLStreamInfo.StreamName][0]
+        self.inputs.clear_stream_buffer_data(EventMarkerLSLStreamInfo.StreamName)
+
+        # state shift
+        for event_marker in event_markers.T:
+            # print(f"working on event marker {event_marker}")
+            block_marker = event_marker[illumiReadSwypeConfig.EventMarkerLSLStreamInfo.BlockChannelIndex]
+            state_marker = event_marker[illumiReadSwypeConfig.EventMarkerLSLStreamInfo.ExperimentStateChannelIndex]
+            user_inputs_marker = event_marker[illumiReadSwypeConfig.EventMarkerLSLStreamInfo.UserInputsChannelIndex]
+
+            # ignore the block_marker <0 and state_marker <0 those means exit the current state
+            if block_marker:
+                if block_marker > 0:
+                    self.enter_block(block_marker)
+                    print(self.currentBlock)
+                else:
+                    self.exit_block(block_marker)
+                    print("Exit Current Block")
+
+            if state_marker:
+                if state_marker > 0:
+                    self.enter_state(state_marker)
+                    print(self.currentExperimentState)
+                else:
+                    self.exit_state(state_marker)
+                    print("Exit Current State")
+
+            # if state_marker and state_marker > 0:  # evoke state change
+            #     self.enter_state(state_marker)
+            #     print(self.currentExperimentState)
+            #     if state_marker == illumiReadSwypeConfig.ExperimentState.KeyboardDewellTimeState or \
+            #             state_marker == illumiReadSwypeConfig.ExperimentState.KeyboardClickState or \
[…]
