### 2023-11-15 b8c307b
move AOI Augmentation
