+++ physiolabxr/scripting/illumiRead/AOIAugmentationScript/AOIAugmentationScript.py
@@ -12,13 +12,13 @@ import torch
 from pylsl import StreamInfo, StreamOutlet, cf_float32
 
 
-from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeData, GazeFilterFixationDetectionIVT, \
+from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeData, GazeFilterFixationDetectionIVT, \
     tobii_gaze_on_display_area_to_image_matrix_index, GazeType, gaze_point_on_image_valid
-from physiolabxr.scripting.AOIAugmentationScript.IntegrateAttention import integrate_attention
+from physiolabxr.scripting.illumiRead.AOIAugmentationScript.IntegrateAttention import integrate_attention
 from physiolabxr.scripting.RenaScript import RenaScript
-from physiolabxr.scripting.AOIAugmentationScript import AOIAugmentationConfig
-from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationUtils import *
-from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationConfig import EventMarkerLSLStreamInfo, GazeDataLSLStreamInfo
+from physiolabxr.scripting.illumiRead.AOIAugmentationScript import AOIAugmentationConfig
+from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationUtils import *
+from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationConfig import EventMarkerLSLStreamInfo, GazeDataLSLStreamInfo
 import torch
 
 
