+++ physiolabxr/scripting/illumiRead/AOIAugmentationScript/version_archive/AOIAugmentationScript.py
@@ -1,13 +1,9 @@
-import time
 from collections import deque
 
-from pylsl import StreamInfo, StreamOutlet, cf_float32
-
-from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeData, GazeFilterFixationDetectionIVT, \
-    tobii_gaze_on_display_area_to_image_matrix_index, GazeType, gaze_point_on_image_valid
+from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeFilterFixationDetectionIVT
 from physiolabxr.scripting.RenaScript import RenaScript
 from physiolabxr.scripting.AOIAugmentationScript import AOIAugmentationConfig
-from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationUtils import *
+from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationUtils import *
 from physiolabxr.scripting.AOIAugmentationScript.AOIAugmentationConfig import EventMarkerLSLStreamInfo, GazeDataLSLStreamInfo
 import torch
 
