### 2023-11-15 faae2b0
update VarjoGazeConfig.py
