+++ physiolabxr/scripting/illumiRead/utils/VarjoEyeTrackingUtils/VarjoGazeConfig.py
@@ -1,0 +1,129 @@
+from enum import Enum
+import numpy as np
+
+
+class IndexClass(int, Enum):
+    pass
+
+
+class VarjoLSLChannelInfo(IndexClass):
+    # Gaze data frame number
+    GazeDataFrameNumber = 0
+
+    # Gaze data capture time (nanoseconds)
+    GazeCaptureTime = 1
+
+    # Log time (milliseconds)
+    LogTime = 2
+
+    # HMD
+    HMDLocalPositionX = 3
+    HMDLocalPositionY = 4
+    HMDLocalPositionZ = 5
+
+    HMDLocalRotationX = 6
+    HMDLocalRotationY = 7
+    HMDLocalRotationZ = 8
+
+    # Combined gaze
+    CombinedGazeValid = 9
+
+    CombinedGazeForwardX = 10
+    CombinedGazeForwardY = 11
+    CombinedGazeForwardZ = 12
+
+    CombinedGazeOriginX = 13
+    CombinedGazeOriginY = 14
+    CombinedGazeOriginZ = 15
+
+    # IDP
+    InterPupillaryDistanceInMM = 16
+
+    # Left eye
+    LeftEyeGazeValid = 17
+
+    LeftGazeForwardX = 18
+    LeftGazeForwardY = 19
+    LeftGazeForwardZ = 20
+
+    LeftGazeOriginX = 21
+    LeftGazeOriginY = 22
+    LeftGazeOriginZ = 23
+
+    LeftPupilIrisDiameterRatio = 24
+    LeftPupilDiameterInMM = 25
+    LeftIrisDiameterInMM = 26
+
+    # Right eye
+    RightEyeGazeValid = 27
+
+    RightGazeForwardX = 28
+    RightGazeForwardY = 29
+    RightGazeForwardZ = 30
+
+    RightGazeOriginX = 31
+    RightGazeOriginY = 32
+    RightGazeOriginZ = 33
+
+    RightPupilIrisDiameterRatio = 34
+    RightPupilDiameterInMM = 35
+    RightIrisDiameterInMM = 36
+
+    # focus
+    FocusDistance = 37
+    FocusStability = 38
+
+    # CombinedGazeRayScreenDirectionX = 0
+    # CombinedGazeRayScreenDirectionY = 1
+    # CombinedGazeRayScreenDirectionZ = 2
+    # CombinedGazeRayScreenOriginX = 3
+    # CombinedGazeRayScreenOriginY = 4
+    # CombinedGazeRayScreenOriginZ = 5
+    # CombinedGazeRayScreenOriginValid = 6
+    # LeftGazeOriginInTrackBoxCoordinatesX = 7
+    # LeftGazeOriginInTrackBoxCoordinatesY = 8
+    # LeftGazeOriginInTrackBoxCoordinatesZ = 9
+    # LeftGazeOriginInUserCoordinatesX = 10
+    # LeftGazeOriginInUserCoordinatesY = 11
+    # LeftGazeOriginInUserCoordinatesZ = 12
+    # LeftGazeOriginValid = 13
+    # LeftGazePointInUserCoordinatesX = 14
+    # LeftGazePointInUserCoordinatesY = 15
+    # LeftGazePointInUserCoordinatesZ = 16
+    # LeftGazePointOnDisplayAreaX = 17
+    # LeftGazePointOnDisplayAreaY = 18
+    # LeftGazePointValid = 19
+    # LeftGazeRayScreenDirectionX = 20
+    # LeftGazeRayScreenDirectionY = 21
+    # LeftGazeRayScreenDirectionZ = 22
+    # LeftGazeRayScreenOriginX = 23
+    # LeftGazeRayScreenOriginY = 24
[…]
