### 2023-11-16 #5854 @mstimberg
> Thank you, @nastaran62, for your thorough review, really appreciate it! We are working through the items you brought up and will get back to you by the end of the coming week.

Hi @ApocalyVec, could you give us an update with respect to the points @nastaran62 raised? Thanks!
