+++ physiolabxr/scripting/Examples/TargetIdentification.py
@@ -1,3 +1,4 @@
+import pickle
 import warnings
 from enum import Enum
 
@@ -16,8 +17,8 @@ event_marker_block_id_index = 1
 item_marker_dtn_index = 0
 item_marker_gaze_intersected_index = 5
 n_chan_per_item = 10
-subimage_size = 64, 64
-# subimage_size = 128, 128
+# subimage_size = 64, 64
+subimage_size = 128, 128
 # capture_size = 1920, 1080
 capture_size = 720, 480
 
@@ -41,10 +42,11 @@ class TargetIdentification(RenaScript):
         Please do not edit this function
         """
         super().__init__(*args, **kwargs)
-        self.images = []
         self.cur_state = States.idle
         self.next_state = States.idle
         self.cur_block_id = None
+        self.image_data = []
+        self.image_labels = []
 
     # Start will be called once when the run button is hit.
     def init(self):
@@ -114,17 +116,18 @@ class TargetIdentification(RenaScript):
 
         item_markers = self.inputs[item_marker_stream_name][0]
         item_markers_times = self.inputs[item_marker_stream_name][1]
-        item_dtns = item_markers[item_marker_dtn_index::n_chan_per_item][:, 0]  # TODO verify dtn columns are the same across the times
+        item_dtns = item_markers[item_marker_dtn_index::n_chan_per_item]  # TODO verify dtn columns are the same across the times
 
         item_gaze_intersects = item_markers[item_marker_gaze_intersected_index::n_chan_per_item]
         item_gaze_intersects = np.diff(item_gaze_intersects, axis=1, prepend=0)
         gaze_onsets = np.argwhere(item_gaze_intersects == 1)  # first column is the item index, second column is the time index
+        gaze_dtn = [item_dtns[x[0], x[1]] for x in gaze_onsets]
+
         gaze_offsets = np.argwhere(item_gaze_intersects == -1)
 
         if len(gaze_onsets) == 0:
             warnings.warn("get_gaze_intersected_images: no gaze intersect found, returning")
             return
-        gaze_dtn = item_dtns[gaze_onsets[:, 0]]
         gaze_times = item_markers_times[gaze_onsets[:, 1]]
 
         # get the gaze positions
@@ -147,11 +150,14 @@ class TargetIdentification(RenaScript):
             plt.show()
 
         # grab the subimages
-        for i, (g_pos, cap) in enumerate(zip(gaze_pos, captures)):
+        for i, (g_pos, cap, label) in enumerate(zip(gaze_pos, captures, gaze_dtn)):
             subimage = cap[int(g_pos[1] - subimage_size[0]/2):int(g_pos[1] + subimage_size[0]/2),
                                   int(g_pos[0] - subimage_size[1]/2):int(g_pos[0] + subimage_size[1]/2), :]
             # subimage = cap[int(g_pos[0]):int(g_pos[0] + subimage_size[0]),
             #            int(g_pos[1] - subimage_size[1] / 2):int(g_pos[1] + subimage_size[1] / 2), :]
-            self.images.append(subimage)
-            plt.imsave(f'{i}.png', subimage)
+            self.image_data.append(subimage)
+            self.image_labels.append(label)
+            plt.imsave(f'gaze{i}.png', subimage)
 
+        pickle.dump(self.image_data, open('image_data.p', 'wb'))
+        pickle.dump(self.image_labels, open('image_labels.p', 'wb'))
