### 2023-11-17 #5854 @HaowenWeiJohn
> > Thank you, @nastaran62, for your thorough review, really appreciate it! We are working through the items you brought up and will get back to you by the end of the coming week.
>
> Hi @ApocalyVec, could you give us an update with respect to the points @nastaran62 raised? Thanks!

Hi @nastaran62,

Thank you for your comments, and we are actively working on the updates. We are in the process of writing documents for the items mentioned [above](https://github.com/openjournals/joss-reviews/issues/5854#issuecomment-1778892975). We will update the paper according to your suggestions.

All the changes will be completed within one and a half weeks, and we will get back to you.

Thank you very much.

Haowen 'John' Wei & PhysioLabXR Team
