### 2023-11-17 87dc549
use re_ prefix for replay stream to avoid duplicate name with the original stream
