+++ physiolabxr/utils/buffers.py
@@ -92,7 +92,9 @@ class DataBuffer():
         self.stream_name_buffer_sizes[stream_name] = size
 
     def _update_buffer(self, stream_name, frames, timestamps):
-
+        """
+        frames: channels x time
+        """
         if stream_name not in self.buffer.keys():
             self.buffer[stream_name] = [np.empty(shape=(frames.shape[0], 0), dtype=frames.dtype),
                                       np.empty(shape=(0,))]  # data first, timestamps second
