### 2023-11-18 3cf2953
unify recording finish dialog; update UI to the recording finish dialog
