+++ physiolabxr/utils/buffers.py
@@ -96,7 +96,7 @@ class DataBuffer():
         frames: channels x time
         """
         if stream_name not in self.buffer.keys():
-            self.buffer[stream_name] = [np.empty(shape=(frames.shape[0], 0), dtype=frames.dtype),
+            self.buffer[stream_name] = [np.empty(shape=(*frames.shape[:-1], 0), dtype=frames.dtype),
                                       np.empty(shape=(0,))]  # data first, timestamps second
 
         # check if the new frame's number of channels matches that of the buffered data
