### 2023-11-21 0f4df69
added VGGTargetId
Added file VGGTargetIdentificationTest.py and recorded and saved some VS data and labels as pickle files
