+++ physiolabxr/examples/WriteYourOwnDataSourceExamples/LSLExampleOutlet.py
@@ -1,16 +1,20 @@
 import time
-import numpy as np
 from pylsl import StreamInfo, StreamOutlet, local_clock
 from random import random as rand
 
 # create a new stream info and outlet
-n_channels = 8
 
+stream_name = 'python_lsl_my_stream_name'
+
+n_channels = 8
+# using the local_clock() to track elapsed time
 start_time = local_clock()
+# track how many samples we have sent
 sent_samples = 0
+# set the sampling rate to 100 Hz
 nominal_sampling_rate = 100
 
-info = StreamInfo('python_lsl_my_stream_name', 'my_stream_type', n_channels, nominal_sampling_rate, 'float32',
+info = StreamInfo('stream_name', 'my_stream_type', n_channels, nominal_sampling_rate, 'float32',
                   'my_stream_id')
 outlet = StreamOutlet(info)
 
