### 2023-11-23 06ebb2b
clean up xdf save/load by not making a special case for XDF
