### 2023-11-24 04c8e57
fix position for the notification to not obstruct the info bar
