### 2023-11-24 c323838
fixed margin for notification pane
