### 2023-11-24 901882b
add set channels constant offset function to settings widget
