### 2023-11-26 f140f62
update ERP classifier example
