+++ physiolabxr/examples/ERPClassifier.py
@@ -1,0 +1,145 @@
+import time
+
+import mne
+import numpy as np
+from sklearn.model_selection import StratifiedShuffleSplit
+
+from physiolabxr.scripting.RenaScript import RenaScript
+from physiolabxr.scripting.events.utils import get_indices_when
+from physiolabxr.scripting.physio.HDCA import HDCA
+from physiolabxr.scripting.physio.epochs import get_event_locked_data, buffer_event_locked_data, \
+    get_baselined_event_locked_data, visualize_epochs
+from physiolabxr.scripting.physio.interpolation import interpolate_zeros
+from physiolabxr.scripting.physio.preprocess import preprocess_samples_eeg_pupil
+
+
+class ERPClassifier(RenaScript):
+    def __init__(self, *args, **kwargs):
+        """
+        Please do not edit this function
+        """
+        super().__init__(*args, **kwargs)
+
+    # Start will be called once when the run button is hit.
+    def init(self):
+        self.dtn_events = (1, 2)  # 1 is distractor, 2 is target
+
+        self.event_marker_name = 'Example-EventMarker-DTN-Block'
+
+        self.eeg_channels = self.get_stream_info('Example-BioSemi-64Chan', 'ChannelNames')  # List of EEG channels
+        self.eeg_srate = self.get_stream_info('Example-BioSemi-64Chan', 'NominalSamplingRate')  # Sampling rate of the EEG data in Hz
+        self.tmin_eeg = -0.1  # Time before event marker to include in the epoch
+        self.tmax_eeg = 0.8  # Time after event marker to include in the epoch
+
+        self.eeg_baseline_time = 0.1  # Time period since the ERP epoch start to use as baseline
+        self.eeg_erp_length = int((self.tmin_eeg - self.tmax_eeg) * self.eeg_srate)  # Length of the ERP epoch in samples
+
+        self.eeg_midlines = ['Fpz', 'AFz', 'Fz', 'FCz', 'Cz', 'CPz', 'Pz', 'POz', 'Oz', 'Iz']  # List of midline EEG channels
+        self.eeg_picks = {self.eeg_channels.index(channel): channel for channel in self.eeg_midlines}  # Indices of the midline EEG channels, used for visualization
+
+        self.eeg_montage = mne.channels.make_standard_montage('biosemi64')
+
+        self.eye_channels = self.get_stream_info('Example-Eyetracking-Pupil', 'ChannelNames')  # List of eye tracking channels
+        self.eye_srate = self.get_stream_info('Example-Eyetracking-Pupil', 'NominalSamplingRate')  # Sampling rate of the eye tracking data in Hz
+
+        self.tmin_eye = -0.5  # Time before event marker to include in the epoch
+        self.tmax_eye = 3.  # Time after event marker to include in the epoch
+        self.eye_baseline_time = 0.5  # Time period since the ERP epoch start to use as baseline
+        self.eye_erp_length = int((self.tmin_eye - self.tmax_eye) * self.eye_srate)  # Length of the ERP epoch in samples
+
+        self.event_locked_data_buffer = {}  # Dictionary to store event-locked data
+        self.process_next_look = False
+        self.block_end_time = None
+        self.block_count = 0
+        self.start_training_at_block = 1
+
+        self.random_seed = 42
+
+
+    # loop is called <Run Frequency> times per second
+    def loop(self):
+        if self.process_next_look:
+            self.process_next_look = False
+            self.block_count += 1
+            # we need to interpolate the pupil data to remove the blinks when pupil size will be zeroes
+            left_pupil = self.inputs['Example-Eyetracking-Pupil'][0][self.eye_channels.index("Left Pupil Size")]
+            right_pupil = self.inputs['Example-Eyetracking-Pupil'][0][self.eye_channels.index("Right Pupil Size")]
[…]
