+++ physiolabxr/scripting/physio/HDCA.py
@@ -1,0 +1,409 @@
+import os
+
+# analysis parameters ######################################################################################
+import matplotlib.pyplot as plt
+import mne
+import numpy as np
+from imblearn.over_sampling import SMOTE
+from mne.viz import plot_topomap
+from numpy.lib.stride_tricks import sliding_window_view
+from sklearn import metrics
+from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
+from sklearn.linear_model import LogisticRegression
+from sklearn.model_selection import StratifiedShuffleSplit
+from sklearn.preprocessing import LabelEncoder
+
+
+def z_norm_hdca(x, _mean=None, _std=None):
+    if _mean is None or _std is None:
+        _mean = np.mean(x, axis=0, keepdims=True)
+        _std = np.std(x, axis=0, keepdims=True)
+    return (x - _mean) / _std
+
+def compute_forward(x_windowed, y, projection):
+    num_train_trials, num_channels, num_windows, num_timepoints_per_window = x_windowed.shape
+    activation = np.empty((2, num_channels, num_windows, num_timepoints_per_window))
+    for class_index in np.sort(np.unique(y)):  # for test set, in increasing order
+        this_x = x_windowed[y == class_index]
+        this_projection = projection[y == class_index]
+        for j in range(num_windows):
+            this_x_window = this_x[:, :, j, :].reshape(this_x.shape[0], -1).T
+            # z_window = np.array([np.dot(weights_channelWindow[j], this_x[trial_index, :, j, :].reshape(-1)) for trial_index in range(this_x.shape[0])])
+            # z_window = z_window.reshape((-1, 1)) # change to a col vector
+            this_projection_window = this_projection[:, j]
+            a = (np.matmul(this_x_window, this_projection_window) / np.matmul(this_projection_window.T, this_projection_window).item()).reshape((num_channels, num_timepoints_per_window))
+            activation[class_index, :, j] = a
+    return activation
+
+def _train_compute_window_projections(x_train_windowed, x_test_windowed, y_train):
+    weights_channelWindow, projection_train_window_trial, ldas = _train_window_lda(x_train_windowed, y_train)
+    projectionTest_window_trial = _compute_window_lda_projections(x_test_windowed, ldas)
+    return weights_channelWindow, projection_train_window_trial, projectionTest_window_trial, ldas
+
+
+def _compute_window_lda_projections(x_windowed, ldas):
+    num_trials, num_channels, num_windows, num_timepoints_per_window = x_windowed.shape
+    projections = np.empty((num_trials, num_windows))
+    for k in range(num_windows):  # iterate over different windows
+        this_x = x_windowed[:, :, k, :].reshape((num_trials, -1))
+        projections[:, k] = ldas[k].transform(this_x).squeeze(axis=1)
+    return projections
+
+def _train_window_lda(x_train_windowed, y_train):
+    num_train_trials, num_channels, num_windows, num_timepoints_per_window = x_train_windowed.shape
+
+    weights_channelWindow = np.empty((num_windows, num_channels * num_timepoints_per_window))
+    projection_train_window_trial = np.empty((num_train_trials, num_windows))
+
+    ldas = []
+    for k in range(num_windows):  # iterate over different windows
+        this_x_train = x_train_windowed[:, :, k, :].reshape((num_train_trials, -1))
+        lda = LinearDiscriminantAnalysis(solver='svd')
+        projection_train_window_trial[:, k] = lda.fit_transform(this_x_train, y_train).squeeze(axis=1)
+        ldas.append(lda)
[…]
