### 2023-11-26 9ca6359
add user input button invoker to illumiReadSwype condition TODO: realtime decoding
