+++ physiolabxr/scripting/illumiRead/illumiReadSwype/illumiReadSwypeScript.py
@@ -7,14 +7,18 @@ import os
 import pickle
 import sys
 import matplotlib.pyplot as plt
+from itertools import groupby
 from physiolabxr.scripting.RenaScript import RenaScript
 from pylsl import StreamInfo, StreamOutlet, cf_float32
 import torch
 from physiolabxr.scripting.illumiRead.illumiReadSwype import illumiReadSwypeConfig
 from physiolabxr.scripting.illumiRead.illumiReadSwype.illumiReadSwypeConfig import EventMarkerLSLStreamInfo, \
-    GazeDataLSLStreamInfo
+    GazeDataLSLStreamInfo, UserInputLSLStreamInfo
+from physiolabxr.scripting.illumiRead.illumiReadSwype.illumiReadSwypeUtils import illumiReadSwypeUserInput
 from physiolabxr.scripting.illumiRead.utils.VarjoEyeTrackingUtils.VarjoGazeUtils import VarjoGazeData
-from physiolabxr.scripting.illumiRead.utils.gaze_utils.general import GazeFilterFixationDetectionIVT
+from physiolabxr.scripting.illumiRead.utils.gaze_utils.general import GazeFilterFixationDetectionIVT, GazeType
+from physiolabxr.utils.buffers import DataBuffer
+
 
 
 class IllumiReadSwypeScript(RenaScript):
@@ -29,6 +33,12 @@ class IllumiReadSwypeScript(RenaScript):
 
         self.currentBlock: illumiReadSwypeConfig.ExperimentBlock = \
             illumiReadSwypeConfig.ExperimentBlock.InitBlock
+
+        self.illumiReadSwyping = False
+
+        self.data_buffer = DataBuffer()
+        self.gaze_data_sequence = list()
+        self.user_input_sequence = list()
 
         self.device = torch.device('cpu')
 
@@ -48,19 +58,24 @@ class IllumiReadSwypeScript(RenaScript):
     def loop(self):
 
         if (EventMarkerLSLStreamInfo.StreamName not in self.inputs.keys()) or (
-                GazeDataLSLStreamInfo.StreamName not in self.inputs.keys()):  # or GazeDataLSLOutlet.StreamName not in self.inputs.keys():
+                GazeDataLSLStreamInfo.StreamName not in self.inputs.keys()) or (
+                UserInputLSLStreamInfo.StreamName not in self.inputs.keys()): # or GazeDataLSLOutlet.StreamName not in self.inputs.keys():
             return
         # print("process event marker call start")
         self.process_event_markers()
 
-        self.process_gaze_data()
+
+        # self.process_gaze_data()
 
         # gaze callback
+
+        self.state_callbacks()
+
         # if self.currentExperimentState == illumiReadSwypeConfig.ExperimentState.KeyboardDewellTimeState or \
         #         self.currentExperimentState == illumiReadSwypeConfig.ExperimentState.KeyboardClickState or \
         #         self.currentExperimentState == illumiReadSwypeConfig.ExperimentState.KeyboardIllumiReadSwypeState or \
         #         self.currentExperimentState == illumiReadSwypeConfig.ExperimentState.KeyboardFreeSwitchState:
-        #     self.process_gaze_data()
+            # self.process_gaze_data()
 
     def cleanup(self):
         print('Cleanup function is called')
@@ -81,6 +96,8 @@ class IllumiReadSwypeScript(RenaScript):
                 if block_marker > 0:
                     self.enter_block(block_marker)
                     print(self.currentBlock)
+
+
                 else:
                     self.exit_block(block_marker)
                     print("Exit Current Block")
@@ -88,6 +105,9 @@ class IllumiReadSwypeScript(RenaScript):
             if state_marker:
                 if state_marker > 0:
[…]
