### 2023-11-26 7bf4edb
use different resample method in ERPClassifier
