### 2023-11-26 dd73056
add get_stream_in_time_range and get_stream_in_index_range
