### 2023-11-26 f763c83
add get_stream_in_time_range and get_stream_in_index_range
