+++ physiolabxr/utils/buffers.py
@@ -205,6 +205,35 @@ class DataBuffer():
     def get_data(self, stream_name):
         return self.buffer[stream_name][0]
 
+    def get_stream_in_time_range(self, stream_name, start_time, end_time):
+
+        # start time must be smaller than end time
+        if start_time > end_time:
+            raise ValueError('start_time must be smaller than end_time')
+
+        stream_data = self.buffer[stream_name][0]
+        stream_timestamps = self.buffer[stream_name][1]
+
+        start_index = np.searchsorted(stream_timestamps, [start_time], side='right')
+        end_index = np.searchsorted(stream_timestamps, [end_time], side='left')
+
+        return [stream_data[:, start_index:end_index+1], stream_timestamps[start_index:end_index+1]]
+
+    def get_stream_in_index_range(self, stream_name, start_index, end_index):
+
+        if start_index < 0 or end_index < 0:
+            raise ValueError('start_index and end_index must be positive')
+
+        if start_index > end_index:
+            raise ValueError('start_index must be smaller than end_index')
+
+        stream_data = self.buffer[stream_name][0]
+        stream_timestamps = self.buffer[stream_name][1]
+
+        return [stream_data[:, start_index:end_index+1], stream_timestamps[start_index:end_index+1]]
+
+
+
     def get_timestamps(self, stream_name):
         return self.buffer[stream_name][1]
 
