### 2023-11-27 999393f
remove connect startStopBtn callback from BaseStreamWidget; child classes just need to override start_stop_stream_btn_clicked from BaseStreamWidget
