### 2023-11-27 78ef9f7
add waiting icon while waiting for stream to become available
