### 2023-11-27 cdb811d
add option for automatically start replay after replay started
