+++ physiolabxr/sub_process/ReplayServer.py
@@ -116,7 +116,7 @@ class ReplayServer(threading.Thread):
                         stream_info[stream_name]['srate'] = stream_info[stream_name]['n_timepoints'] / (timestamps[-1] - timestamps[0])
                         stream_info[stream_name]['data_type'] = str(data.dtype)
                     self.send_string(json.dumps(stream_info))
-                elif command == shared.GO_AHEAD_COMMAND:
+                elif command == shared.SETUP_STREAM_COMMAND:
                     # go ahead and open the streams
                     # outlets are not created in setup before receiving the go-ahead from the main process because the
                     # main process need to check if there're duplicate stream names with the streams being replayed
@@ -143,8 +143,18 @@ class ReplayServer(threading.Thread):
                         continue
                     self.send_string(shared.SUCCESS_INFO)
                     self.send(np.array([self.start_time, self.end_time, self.total_time, self.virtual_clock_offset]))  # send the timing info
-                    print(f"replay started for streams: {list(self.stream_data)}")
-                    self.is_replaying = True  # this is the only entry point of the replay loop
+
+                    # hold on for the start replay command
+                    # replay tab needs to tell MainWindow to add the stream widgets, it may take a while
+                    wait_start_time = time.time()
+                    command = self.recv_string(is_block=True)
+                    if command == shared.START_REPLAY_COMMAND:
+                        print(f"After waiting main process for {time.time() - wait_start_time} seconds, replay started for streams: {list(self.stream_data)}")
+                        self.is_replaying = True  # this is the only entry point of the replay loop
+                    else:
+                        print(f"ReplayServer: unexpected command received: {command}, resetting replay")
+                        self.reset_replay()
+                        continue
                 elif command == shared.PERFORMANCE_REQUEST_COMMAND:
                     self.send(self.get_average_loop_time())
                 elif command == shared.TERMINATE_COMMAND:
