### 2023-11-30 c5fac36
remove prints from Dtest.py
