### 2023-12-04 #5854 @mstimberg
> Thank you for your comments, and we are actively working on the updates. We are in the process of writing documents for the items mentioned [above](https://github.com/openjournals/joss-reviews/issues/5854#issuecomment-1778892975). We will update the paper according to your suggestions.
>
> All the changes will be completed within one and a half weeks, and we will get back to you.

:wave: @HaowenWeiJohn can you give us an update with respect to this?
