### 2023-12-04 b986715
update vgg object identification
