### 2023-12-04 7b088e0
"Implement spelling correction and swyping feature
The changes include implementation of spelling correction functionality using neuspell in 'neuspell_utils.py'. Also, the 'illumiReadSwypeUtils.py' script has been updated with a swyping functionality to predict words based on swyped characters. The mapping of characters in 'illumiReadSwypeConfig.py' are enhanced to facilitate the above additions. Further, 'illumiReadSwypeScript.py' has been refactored to include spelling correction and streams associated for data preservation. These changes are aimed to improve user input prediction and to pro
[…]
