+++ physiolabxr/scripting/illumiRead/illumiReadSwype/illumiReadSwypeUtils.py
@@ -1,6 +1,7 @@
 import numpy as np
 
 from physiolabxr.scripting.illumiRead.illumiReadSwype import illumiReadSwypeConfig
+from physiolabxr.scripting.illumiRead.illumiReadSwype.illumiReadSwypeConfig import KeyIndexIDDict
 
 
 class illumiReadSwypeUserInput:
@@ -42,10 +43,31 @@ def word_candidate_list_to_lvt(words_list, max_length=1024):
     Convert a words sequence to a list of integers representing the characters
     """
     lvt = [0]
+    lvt.append(len(words_list))
+    overflow_flag = False
+
+    for word_index, word in enumerate(words_list):
+        word_lvt = []
+        word_lvt.append(word_index)
+        word_lvt.append(len(word))
+        for char in word:
+            chat_int = KeyIndexIDDict[char.upper()]
+            word_lvt.append(chat_int)
+
+        lvt += word_lvt
+
+        if len(lvt) >= max_length:
+            overflow_flag = True
+            break
+
+    if not overflow_flag:
+        lvt += [0] * (max_length - len(lvt))
+    else:
+        lvt[0] = 1
+
+    return lvt, overflow_flag
 
 
+if __name__ == '__main__':
 
-
-
-
-    pass
+    word_candidate_list_to_lvt(['hello', 'world', 'this'])
