+++ physiolabxr/scripting/illumiRead/illumiReadSwype/illumiReadSwypeScript.py
@@ -18,8 +18,8 @@ from physiolabxr.scripting.illumiRead.illumiReadSwype.illumiReadSwypeUtils impor
     word_candidate_list_to_lvt
 from physiolabxr.scripting.illumiRead.utils.VarjoEyeTrackingUtils.VarjoGazeUtils import VarjoGazeData
 from physiolabxr.scripting.illumiRead.utils.gaze_utils.general import GazeFilterFixationDetectionIVT, GazeType
+from physiolabxr.scripting.illumiRead.utils.language_utils.neuspell_utils import SpellCorrector
 from physiolabxr.utils.buffers import DataBuffer
-
 
 
 class IllumiReadSwypeScript(RenaScript):
@@ -47,9 +47,21 @@ class IllumiReadSwypeScript(RenaScript):
 
         self.process_gaze_data_time_buffer = deque(maxlen=1000)
 
-
         self.ivt_filter = GazeFilterFixationDetectionIVT(angular_speed_threshold_degree=100)
 
+        # spelling correction
+        self.spell_corrector = SpellCorrector()
+        self.spell_corrector.correct_string("WHAT")
+
+        # create stream outlets
+        illumireadswype_keyboard_suggestion_strip_lsl_stream_info = StreamInfo(
+            illumiReadSwypeConfig.illumiReadSwypeKeyboardSuggestionStripLSLStreamInfo.StreamName,
+            illumiReadSwypeConfig.illumiReadSwypeKeyboardSuggestionStripLSLStreamInfo.StreamType,
+            illumiReadSwypeConfig.illumiReadSwypeKeyboardSuggestionStripLSLStreamInfo.ChannelNum,
+            illumiReadSwypeConfig.illumiReadSwypeKeyboardSuggestionStripLSLStreamInfo.NominalSamplingRate,
+            channel_format=cf_float32)
+
+        self.illumireadswype_keyboard_suggestion_strip_lsl_outlet = StreamOutlet(illumireadswype_keyboard_suggestion_strip_lsl_stream_info)  # shape: (1024, 1)
 
 
     def init(self):
@@ -60,11 +72,10 @@ class IllumiReadSwypeScript(RenaScript):
 
         if (EventMarkerLSLStreamInfo.StreamName not in self.inputs.keys()) or (
                 GazeDataLSLStreamInfo.StreamName not in self.inputs.keys()) or (
-                UserInputLSLStreamInfo.StreamName not in self.inputs.keys()): # or GazeDataLSLOutlet.StreamName not in self.inputs.keys():
+                UserInputLSLStreamInfo.StreamName not in self.inputs.keys()):  # or GazeDataLSLOutlet.StreamName not in self.inputs.keys():
             return
         # print("process event marker call start")
         self.process_event_markers()
-
 
         # self.process_gaze_data()
 
@@ -76,7 +87,7 @@ class IllumiReadSwypeScript(RenaScript):
         #         self.currentExperimentState == illumiReadSwypeConfig.ExperimentState.KeyboardClickState or \
         #         self.currentExperimentState == illumiReadSwypeConfig.ExperimentState.KeyboardIllumiReadSwypeState or \
         #         self.currentExperimentState == illumiReadSwypeConfig.ExperimentState.KeyboardFreeSwitchState:
-            # self.process_gaze_data()
+        # self.process_gaze_data()
 
     def cleanup(self):
         print('Cleanup function is called')
@@ -125,8 +136,6 @@ class IllumiReadSwypeScript(RenaScript):
             #         print("state change")
 
             # print("process event marker call end")
-
-
 
     def enter_block(self, block_marker):
 
@@ -222,7 +231,6 @@ class IllumiReadSwypeScript(RenaScript):
[…]
