### 2023-12-05 c0e865f
add output directory in target identification
