### 2023-12-05 fdff78c
fix nothing received empty list bug
