### 2023-12-09 9ef68c2
"Add class rebalancing with SMOTE in utils.py"
Imbalanced classes can hinder the efficiency and accuracy of machine learning algorithms. To handle this issue, Synthetic Minority Oversampling Technique (SMOTE) method has been added to the utils.py file. It allows users to rebalance classes in the input data, either on the whole or per individual channel. Import and usage of SMOTE has been also included in a new function called 'rebalance_classes'. The function takes in parameters for the input data, target labels, a boolean parameter to balance by channel or not, and an optional random seed.
