### 2023-12-09 182a2a4
"Refactor ERPClassifier for better readability and maintainability"
Updated ERPClassifier.py to split the training and test prediction process which was previously combined, for better readability and maintainability. Removed an unnecessary SVM classifier that was presenting duplicated code and performance results. Also, omits the detailed print summaries for each block.
The deprecated SVM classifier was moved to a new file, ERPClassifierSVM.py in the deprecated directory, to keep it for reference. It's highly recommended to use the refactored version for future works.
