### 2023-12-12 4798cbf
add none type check in camera image read; not use format keyword in listcomp in OptionsWindowPlotFormatWidget.py; add placeholder item to plot options window to avoid NSRangeException on Mac; add exception bypass for process frames; add placeholder text in settings widget
