### 2023-12-12 e67c4a2
add place hold text for comboboxes throwing NSRangeException
