+++ physiolabxr/utils/ui_utils.py
@@ -181,8 +181,7 @@ def stream_stylesheet(stylesheet_url):
 #     widget.setStyleSheet(AppConfigs()._style_sheets[theme])
 
 def add_presets_to_combobox(combobox: QComboBox):
-    for i in get_all_preset_names():
-        combobox.addItem(i)
+    combobox.addItems([i for i in get_all_preset_names()])
 
 def update_presets_to_combobox(combobox: QComboBox):  # TODO script should also call this when new preset is added
     combobox.clear()
@@ -190,8 +189,7 @@ def update_presets_to_combobox(combobox: QComboBox):  # TODO script should also 
         combobox.addItem(i)
 
 def add_stream_presets_to_combobox(combobox):
-    for i in get_stream_preset_names():
-        combobox.addItem(i)
+    combobox.addItems([i for i in get_stream_preset_names()])
 
 class AnotherWindow(QWidget):
     """
@@ -285,7 +283,7 @@ def validate_script_path(script_path, desired_class: Type) -> bool:
 
 def add_enum_values_to_combobox(combobox: QComboBox, enum: Type[Enum], can_be_selected_in_gui: list=None):
     if can_be_selected_in_gui is None:
-        add_items(combobox, [name for name, member in enum.__members__.items()])
+        combobox.addItems([name for name, member in enum.__members__.items()])
     else:
         for name, member in enum.__members__.items():
             combobox.addItem(name)
@@ -317,18 +315,18 @@ def get_int_from_line_edit(line_edit: QtWidgets.QLineEdit, name=""):
         line_edit.textChanged.connect(set_back)
         raise RenaError(f'{name} must be an integer')
 
-def add_items(combobox: QComboBox, items: Iterable):
-    """
-    call addItems on a combobox
-    remove placeholder if there's any
-    """
-    placeholder_index = combobox.findText(AppConfigs()._placeholder_text)
-    if placeholder_index == -1:
-        warnings.warn(f"combobox {combobox.objectName()} has no placeholder, may subject to NSException.")
-    combobox.addItems(items)
-    # remove the placeholder item from the combobox if it exists
-    if placeholder_index != -1:
-        combobox.removeItem(placeholder_index)
+# def add_items(combobox: QComboBox, items: Iterable):
+#     """
+#     call addItems on a combobox
+#     remove placeholder if there's any
+#     """
+#     placeholder_index = combobox.findText(AppConfigs()._placeholder_text)
+#     if placeholder_index == -1:
+#         warnings.warn(f"combobox {combobox.objectName()} has no placeholder, may subject to NSException.")
+#     combobox.addItems(items)
+#     # remove the placeholder item from the combobox if it exists
+#     if placeholder_index != -1:
+#         combobox.removeItem(placeholder_index)
 
 
 
