### 2023-12-13 932d2d1
update classifier example, add example experiment preset
