+++ physiolabxr/examples/ERPClassifier.py
@@ -1,5 +1,6 @@
 import time
 
+import cv2
 import numpy as np
 import scipy
 from sklearn.model_selection import StratifiedShuffleSplit
@@ -28,6 +29,10 @@ class ERPClassifier(RenaScript):
         self.dtn_events = (1, 2)  # 1 is distractor, 2 is target
 
         self.event_marker_name = 'Example-EventMarker-DTN-Block'
+        self.video_output_name = 'Example-Video-Output'
+        self.video_input_name = 'Example-Video'
+        self.video_shape = (400, 400, 3)  # the shape of the video stream, because all inputs are flattened as they comes in, we need to reshape the video frames to be able to put shapes on them
+        self.last_dtn_time = 0
 
         self.eeg_channels = self.get_stream_info('Example-BioSemi-64Chan', 'ChannelNames')  # List of EEG channels
         self.eeg_srate = self.get_stream_info('Example-BioSemi-64Chan', 'NominalSamplingRate')  # Sampling rate of the EEG data in Hz
@@ -56,8 +61,12 @@ class ERPClassifier(RenaScript):
 
         self.random_seed = 42
 
+        self.model = None
+        self.event_indicators = []
+        self.prediction_indicators = []
+        self.indicator_radius = int(0.5 * 0.9 * self.video_shape[0] // 30)  # 0.5 for the radius, 0.9 for the margins, 30 for the number items per block
+        self.indicator_separation = int(self.video_shape[0] // 30)
 
-    # loop is called <Run Frequency> times per second
     def loop(self):
         if self.process_next_look:
             self.process_next_look = False
@@ -81,7 +90,6 @@ class ERPClassifier(RenaScript):
                                                                        tmax={'eeg': self.tmax_eeg, 'eye': self.tmax_eye},
                                                                        srate={'eeg': self.eeg_srate, 'eye': self.eye_srate},
                                                                        return_last_event_time=True, verbose=1)
-
 
             self.inputs.clear_up_to(self.block_end_time)  # Clear the input buffer up to the last event time to avoid processing duplicate data
             self.event_locked_data_buffer = buffer_event_locked_data(event_locked_data, self.event_locked_data_buffer)  # Buffer the event-locked data for further processing
@@ -137,12 +145,12 @@ class ERPClassifier(RenaScript):
 
                 # hdca classifier
 
-                hdca_model = HDCA(target_names)
-                roc_auc_combined_train, roc_auc_eeg_train, roc_auc_pupil_train = hdca_model.fit(x_eeg_train, x_eeg_pca_ica_train, x_pupil_train, y_train, num_folds=1, is_plots=True, exg_srate=self.eeg_srate, notes=f"Block ID {self.block_count}", verbose=0, random_seed=self.random_seed)  # give the original eeg data, no need to apply HDCA again
-                y_pred_train, roc_auc_eeg_pupil_test, roc_auc_eeg_test, roc_auc_pupil_test = hdca_model.eval(x_eeg_train, x_eeg_pca_ica_train, x_pupil_train, y_train, notes=f"Block ID {self.block_count}")
-                y_pred_test, roc_auc_eeg_pupil_test, roc_auc_eeg_test, roc_auc_pupil_test = hdca_model.eval(x_eeg_test, x_eeg_pca_ica_test, x_pupil_test, y_test, notes=f"Block ID {self.block_count}")
-                # report the results
+                self.model = HDCA(target_names)
[…]
