### 2023-12-13 d581882
Add replay auto-set-to-zmq when # channels exceeds a certain value
