+++ physiolabxr/configs/configs.py
@@ -84,6 +84,8 @@ class AppConfigs(metaclass=Singleton):
 
     # replay configs
     start_streams_on_replay: bool = True
+    auto_select_zmq_if_exceed_n_channels: bool = True
+    auto_select_zmq_n_channels: int = int(2 ** 9)
 
     # recording configs
     recording_file_format: RecordingFileFormat = RecordingFileFormat.dats
