### 2023-12-13 a9effb6
add dtn indicators in video output
