### 2023-12-13 67d9e61
add single epoch prediction to multi-modal classifier example
