+++ physiolabxr/examples/ERPClassifier.py
@@ -13,7 +13,7 @@ from physiolabxr.scripting.physio.HDCA import HDCA
 from physiolabxr.scripting.physio.epochs import get_event_locked_data, buffer_event_locked_data, \
     get_baselined_event_locked_data, visualize_epochs
 from physiolabxr.scripting.physio.interpolation import interpolate_zeros
-from physiolabxr.scripting.physio.preprocess import preprocess_samples_eeg_pupil
+from physiolabxr.scripting.physio.preprocess import preprocess_samples_eeg_pupil, Preprocessor
 from physiolabxr.scripting.physio.utils import rebalance_classes
 
 
@@ -57,107 +57,56 @@ class ERPClassifier(RenaScript):
         self.process_next_look = False
         self.block_end_time = None
         self.block_count = 0
-        self.start_training_at_block = 1
 
         self.random_seed = 42
 
         self.model = None
-        self.event_indicators = []
-        self.prediction_indicators = []
+        self.eeg_pca = None
+        self.eeg_ica = None
+        self.event_indicators = {}
         self.indicator_radius = int(0.5 * 0.9 * self.video_shape[0] // 30)  # 0.5 for the radius, 0.9 for the margins, 30 for the number items per block
         self.indicator_separation = int(self.video_shape[0] // 30)
 
     def loop(self):
         # wait until we have received the last pupil block
-        if self.process_next_look and self.inputs['Example-Eyetracking-Pupil'][1][-1] - 3 > self.block_end_time:
+        # the following line gets the last pupil time, this is used to check if we have received the last pupil epoch, which is 3 seconds after the event
+        last_pupil_time = self.inputs['Example-Eyetracking-Pupil'][1][-1] if 'Example-Eyetracking-Pupil' in self.inputs.keys() else 0
+        if self.process_next_look and last_pupil_time - 3 > self.block_end_time:
             self.process_next_look = False
             self.block_count += 1
-            # we need to interpolate the pupil data to remove the blinks when pupil size will be zeroes
-            left_pupil = self.inputs['Example-Eyetracking-Pupil'][0][self.eye_channels.index("Left Pupil Size")]
-            right_pupil = self.inputs['Example-Eyetracking-Pupil'][0][self.eye_channels.index("Right Pupil Size")]
-            self.inputs['Example-Eyetracking-Pupil'][0][self.eye_channels.index("Left Pupil Size")] = interpolate_zeros(left_pupil)
-            self.inputs['Example-Eyetracking-Pupil'][0][self.eye_channels.index("Right Pupil Size")] = interpolate_zeros(right_pupil)
 
-            # filter the eeg data
-            # self.inputs['Example-BioSemi-64Chan'][0] = mne.filter.filter_data(self.inputs['Example-BioSemi-64Chan'][0], sfreq=self.eeg_srate, l_freq=1, h_freq=50, n_jobs=1)
-
-            # get the event-locked EEG data
-            # get the event-locked eeg and eye (pupil) data
-            event_locked_data, last_event_time = get_event_locked_data(event_marker=self.inputs[self.event_marker_name],
-                                                                       data={'eeg': self.inputs['Example-BioSemi-64Chan'],
-                                                                             'eye': self.inputs['Example-Eyetracking-Pupil']},
-                                                                       events_of_interest=self.dtn_events,
-                                                                       tmin={'eeg': self.tmin_eeg, 'eye': self.tmin_eye},
-                                                                       tmax={'eeg': self.tmax_eeg, 'eye': self.tmax_eye},
[…]
