+++ physiolabxr/examples/ERPClassifier.py
@@ -1,10 +1,7 @@
-import time
-
 import cv2
 import numpy as np
 import scipy
 from sklearn.model_selection import StratifiedShuffleSplit
-from sklearn import svm
 from sklearn.metrics import classification_report
 
 from physiolabxr.scripting.RenaScript import RenaScript
@@ -13,7 +10,7 @@ from physiolabxr.scripting.physio.HDCA import HDCA
 from physiolabxr.scripting.physio.epochs import get_event_locked_data, buffer_event_locked_data, \
     get_baselined_event_locked_data, visualize_epochs
 from physiolabxr.scripting.physio.interpolation import interpolate_zeros
-from physiolabxr.scripting.physio.preprocess import preprocess_samples_eeg_pupil, Preprocessor
+from physiolabxr.scripting.physio.preprocess import preprocess_samples_eeg_pupil
 from physiolabxr.scripting.physio.utils import rebalance_classes
 
 
@@ -63,9 +60,10 @@ class ERPClassifier(RenaScript):
         self.model = None
         self.eeg_pca = None
         self.eeg_ica = None
-        self.event_indicators = {}
+        self.event_indicators = []
         self.indicator_radius = int(0.5 * 0.9 * self.video_shape[0] // 30)  # 0.5 for the radius, 0.9 for the margins, 30 for the number items per block
-        self.indicator_separation = int(self.video_shape[0] // 30)
+        self.indicator_separation = int(1.1 * self.video_shape[0] // 30)
+        self.horizontal_offset = 5
 
     def loop(self):
         # wait until we have received the last pupil block
@@ -75,14 +73,13 @@ class ERPClassifier(RenaScript):
             self.process_next_look = False
             self.block_count += 1
 
-            event_locked_data, last_event_time, baselined_eeg_epochs, baselined_resampled_pupil_epochs, x_eeg_znormed, x_eeg_pca_ica, x_pupil_znormed, y\
+            last_event_time, baselined_eeg_epochs, baselined_resampled_pupil_epochs, x_eeg_znormed, x_eeg_pca_ica, x_pupil_znormed, y\
                 = self.process_data(self.inputs[self.event_marker_name], is_rebalance=True)
 
             self.inputs.clear_up_to(self.block_end_time)  # Clear the input buffer up to the last event time to avoid processing duplicate data
-            self.event_locked_data_buffer = buffer_event_locked_data(event_locked_data, self.event_locked_data_buffer)  # Buffer the event-locked data for further processing
 
-            visualize_epochs(baselined_eeg_epochs, picks=self.eeg_picks)
-            visualize_epochs(baselined_resampled_pupil_epochs)
+            visualize_epochs(baselined_eeg_epochs, picks=self.eeg_picks, tmin=self.tmin_eeg, tmax=self.tmax_eeg, title='EEG')
+            visualize_epochs(baselined_resampled_pupil_epochs, tmin=self.tmin_eye, tmax=self.tmax_eye, title='Pupil Size')
 
             # split the data into train and test, we only do one split as we are not doing cross validation here
             skf = StratifiedShuffleSplit(n_splits=1, random_state=self.random_seed, test_size=0.2)
@@ -129,27 +126,33 @@ class ERPClassifier(RenaScript):
                 # add the dtn events to the event indicators, so they can be visualized on the video
                 new_dtn_events = [{'time': t, 'type': event_type} for t, event_type in zip(dtns_times, found_dtns[:, 0])]
                 self.event_indicators += new_dtn_events
-                print(f"Found {len(dtn_indices)} DTN events, event indicator length {len(self.event_indicators)}, content: {self.event_indicators}")
[…]
