### 2023-12-15 6054010
save classifier figures to local disk
