+++ physiolabxr/examples/ERPClassifier.py
@@ -1,3 +1,5 @@
+import os
+
 import cv2
 import numpy as np
 import scipy
@@ -12,6 +14,7 @@ from physiolabxr.scripting.physio.epochs import get_event_locked_data, buffer_ev
 from physiolabxr.scripting.physio.interpolation import interpolate_zeros
 from physiolabxr.scripting.physio.preprocess import preprocess_samples_eeg_pupil
 from physiolabxr.scripting.physio.utils import rebalance_classes
+from physiolabxr.utils.data_utils import get_date_string
 
 
 class ERPClassifier(RenaScript):
@@ -65,6 +68,12 @@ class ERPClassifier(RenaScript):
         self.indicator_separation = int(1.1 * self.video_shape[0] // 30)
         self.horizontal_offset = 5
 
+        # create a 'figure' folder to save the figures generated by the script
+        # get a date string to be used as the folder name
+        self.figure_folder = os.path.join('figures', get_date_string())
+        os.makedirs(self.figure_folder, exist_ok=True)
+        print(f"Figures will be saved to {self.figure_folder}")
+
     def loop(self):
         # wait until we have received the last pupil block
         # the following line gets the last pupil time, this is used to check if we have received the last pupil epoch, which is 3 seconds after the event
@@ -78,8 +87,8 @@ class ERPClassifier(RenaScript):
 
             self.inputs.clear_up_to(self.block_end_time)  # Clear the input buffer up to the last event time to avoid processing duplicate data
 
-            visualize_epochs(baselined_eeg_epochs, picks=self.eeg_picks, tmin=self.tmin_eeg, tmax=self.tmax_eeg, title='EEG')
-            visualize_epochs(baselined_resampled_pupil_epochs, tmin=self.tmin_eye, tmax=self.tmax_eye, title='Pupil Size')
+            visualize_epochs(baselined_eeg_epochs, picks=self.eeg_picks, tmin=self.tmin_eeg, tmax=self.tmax_eeg, title=f'EEG Epochs, block {self.block_count}', out_dir=self.figure_folder)
+            visualize_epochs(baselined_resampled_pupil_epochs, tmin=self.tmin_eye, tmax=self.tmax_eye, title=f'Pupil Size Epochs, block {self.block_count}', out_dir=self.figure_folder)
 
             # split the data into train and test, we only do one split as we are not doing cross validation here
             skf = StratifiedShuffleSplit(n_splits=1, random_state=self.random_seed, test_size=0.2)
@@ -93,7 +102,7 @@ class ERPClassifier(RenaScript):
 
             # hdca classifier
             self.model = HDCA(target_names)
-            roc_auc_combined_train, roc_auc_eeg_train, roc_auc_pupil_train = self.model.fit(x_eeg_znormed_train, x_eeg_pca_ica_train, x_pupil_train, y_train, num_folds=1, is_plots=True, exg_srate=self.eeg_srate, notes=f"Block ID {self.block_count}", verbose=0, random_seed=self.random_seed)  # give the original eeg data, no need to apply HDCA again
+            roc_auc_combined_train, roc_auc_eeg_train, roc_auc_pupil_train = self.model.fit(x_eeg_znormed_train, x_eeg_pca_ica_train, x_pupil_train, y_train, num_folds=1, is_plots=True, exg_srate=self.eeg_srate, notes=f"Block ID {self.block_count}", verbose=0, random_seed=self.random_seed, out_dir=self.figure_folder)  # give the original eeg data, no need to apply HDCA again
             y_pred_train, roc_auc_eeg_pupil_test, roc_auc_eeg_test, roc_auc_pupil_test = self.model.eval(x_eeg_znormed_train, x_eeg_pca_ica_train, x_pupil_train, y_train, notes=f"Block ID {self.block_count}")
[…]
