### 2023-12-15 #5854 @ApocalyVec
> The section "Python API, Write your data source script" doesn't have any content. Also, it seems that there is no solution or explanation for visualizing the data recorded in this way.

We reorganized everything concerning data sources under “[Stream Data](https://physiolabxrdocs.readthedocs.io/en/latest/DataStreamAPI.html)” in the docs. We split the LSL, ZMQ, and script-based data source in to their separate pages:

* [Stream data with ZeroMQ (ZMQ)](https://physiolabxrdocs.readthedocs.io/en/latest/DataStreamAPI/ZMQStream.html)
* [Stream data with Lab Streaming Layer (LSL)](https://physiolabxrdocs.readthedocs.io/en/latest/DataStreamAPI/LSLStream.html)
* [Create data source using Scripting Interface](https://physiolabxrdocs.readthedocs.io/en/latest/DataStreamAPI/ScriptStream.html#)
