### 2023-12-15 #5854 @ApocalyVec
> I can't see any API reference in the documents (No document in the code). API documentation are very helpful in tutorials/examples.

Thank you for the suggestion! We created dedicated API reference pages in our docs, with all the key API functions. The docs will always be up to date with the latest releases. Here are the ones we have currently:
[RNStream](https://physiolabxrdocs.readthedocs.io/en/latest/APIReferences/RNStream.html)
[DATS (Dictionary of Array and Timestamps) File Format](https://physiolabxrdocs.readthedocs.io/en/latest/APIReferences/DATSFileFormat.html)
[Fixation](https://physiolabxrdocs.readthedocs.io/en/latest/APIReferences/FixationAPIRef.html)
[Buffer API Class](https://physiolabxrdocs.readthedocs.io/en/latest/APIReferences/BufferAPIRef.html)
[ERP API Reference](https://physiolabxrdocs.readthedocs.io/en/latest/APIReferences/ERPAPIRef.html)
[Stream_in API Reference](https://physiolabxrdocs.readthedocs.io/en/latest/APIReferences/StreamIn.html)
[get_stream_info API Reference](https://physiolabxrdocs.readthedocs.io/en/latest/APIReferences/get_stream_info_APIRef.html)
