### 2023-12-15 #5854 @ApocalyVec
> This tutorial, "Building ERP Classifier based on EEG and Pupil Size" is not clear.

Sorry about the confusion, this tutorial is fully online now: [Building ERP Classifier based on EEG and Pupil Size](https://physiolabxrdocs.readthedocs.io/en/latest/tutorials/BuildMultiModalERPClassifier.html).  We have intended this tutorial to be a step up from the “[Simple ERP Example](https://physiolabxrdocs.readthedocs.io/en/latest/index.html#get-started-with-a-simple-example)”The main takeaway from this tutorial is a classifier that distinguishes between distractors and targets in a visual oddball task, using ERP signals from both EEG and pupillometry data. The data pipeline includes the integration of EEG data, eye-tracking information, and event marker streams. This setup makes use of PhysioLabXR's replay and scripting features to train and execute the classifier in real-time as the participant performs the visual oddball task.
