### 2023-12-16 #5854 @ApocalyVec
> It's better to keep the figures' captions short and explain them in the main body.

We are following the standard of self-contained figures with long captions (as commonly seen in neuroscience papers). Having not been provided a guideline on captions styles by [the paper docs](https://joss.readthedocs.io/en/latest/submitting.html), we looked into the community and did a quick search: the papers published on JOSS seem to be a mixture of self-contained (e.g., [Lucas et al.](https://joss.theoj.org/papers/10.21105/joss.05872) ) and short captions (e.g., [Silge et al.]([https://joss.theoj.org/papers/10.21105/joss.00037](https://joss.theoj.org/papers/10.21105/joss.00037.pdf)) ).
For now, we opted to keep the captions as they are for our audience. Please let us know if we should rewrite the captions.
