### 2023-12-16 #5854 @ApocalyVec
> Installation
> 1- I installed it with pip. It was done successfully. However, it had some errors in pyaudio installation (Ubuntu OS).

Thank you for pointing this out. We weren’t able to reproduce the error on our end with Ubuntu versions 20.04.6 LTS,  22.04.2 LTS, and 23.10. Could you please provide us with more details on the error you receive, like console logs and the OS version?
