### 2023-12-16 ef4ad3e
add set channels constant offset function to settings widget
