+++ paper/paper.bib
@@ -563,3 +563,16 @@
   publisher={Society of Photo-Optical Instrumentation Engineers},
   doi={10.1117/1.NPh.4.4.041504}
 }
+
+
+@article{saffaryazdi2022octopus,
+  title={Octopus Sensing: A Python library for human behavior studies},
+  author={Saffaryazdi, Nastaran and Gharibnavaz, Aidin and Billinghurst, Mark},
+  journal={Journal of Open Source Software},
+  volume={7},
+  number={71},
+  pages={4045},
+  year={2022}
+}
+
+
