### 2023-12-16 #5854 @HaowenWeiJohn
We would like to thank the reviewers and editors for their comprehensive suggestions and comments.

> 3- In several parts of the paper, the authors claimed high performance and rapid prototyping for developing experiments. I suggest comparing its performance with similar software or libraries like [PsychoPy](https://www.psychopy.org/) and [Octopus Sensing](https://octopus-sensing.nastaran-saffar.me/)
> 4- I can see this software has many features similar to [Octopus Sensing](https://octopus-sensing.nastaran-saffar.me/) (related paper in JOSS: https://joss.theoj.org/papers/10.21105/joss.04045). It may be good to discuss why you chose not to integrate with or extend it.
>

@nastaran62
PsychoPy is a platform designed for experiment creation. In other words, it is a stimulus presentation software with high timing accuracy. Experiments built with PsychoPy can be modified to send event markers that can be used with other data sources like EEG and eye tracking. In contrast, PhysioLabXR is a data hub that takes in these data sources (e.g., including the event marker generated from PsychoPy) and offers functionalities including visualization, recording, replay, DSP. It also has a versatile scripting interface that allows users to do anything with the data such as train ML models and make inferences in real-time (e.g., (the P300 speller tutorial)[https://physiolabxrdocs.readthedocs.io/en/latest/PhysioLabXRP300SpellerDemo.html], (object detection tutorial)[https://physiolabxrdocs.readthedocs.io/en/latest/ObjectDetectionTutorial.html]).

Besides, we have a tutorial [“fNIRS Stroop Experiment with PhysiolabXR and PsychoPy”](https://physiolabxrdocs.readthedocs.io/en/latest/PsychoPy.html) that shows how to integrate PsychoPy with PhysioLabXR.

Thanks for pointing out (Octopus-sensing)[https://joss.theoj.org/papers/10.21105/joss.04045], which wasn’t under our radar at the time of writing the platform. Octopus-sensing indeed has a number of great features such as worker processes for acquiring data and a web-based visualizer. Parts of PhysioLabXR do resemble Octopus-sensing. Nevertheless, PhysioLabXR seeks to offer a different user experience, notable ones include:

* The management of the streams done is through a GUI, abstracting the data workers behind the scene.
* PhysioLabXR also has built-in signal processors (e.g.,  filters) or scripting interface with an integrated GUI.
* Apart from native support for a number of sensors, cameras, audio devices, screen capture, PhysioLabXR additionally supports LSL or ZMQ for data transfer. For non-native devices, similar to the “(adding support for a new device)[https://octopus-sensing.nastaran-saffar.me/development.html#adding-support-for-a-new-device]” in Octopus-sensing, users can use the PhysioLabXR’s scripting interface to create customized data sources, but the scope of the scripting feature is not limited to new data sources, extending beyond a device driver. Its potential lies in the real-time analysis of data streams, with scalable i/o ports, real-time parameter adjustments for fine tuning an experiment (e.g., [real-time fixation detection](https://physiolabxrdocs.readthedocs.io/en/latest/FixationDetection.html) ).

As for the visualization, we were trying to run the octopus-sensing-visualizer but run into some trouble firing up the visualization (please see  [octopus-sensing/issues/70](https://github.com/octopus-sensing/octopus-sensing/issues/70)).
