### 2023-12-16 #5854 @HaowenWeiJohn
> Software:
>> 1-Usually, in EEG-based studies, we need to send a marker or trigger to mark an event on data between start and stop. It seems that PhysioLabXR doesn't support this. or maybe it records markers in a separate file from the sensor data. (In this case, we need a script to match it with data and know when exactly the event happened, for example, visualization or processing). Could you please explain if it is possible to add it to PhysioLabXR, or how do you handle this currently? I also suggest adding it as an issue to the repository to resolve in the future.
>

Thanks for pointing this out! To clarify, PhysioLabXR is not currently designed as a stimulus presentation software (SPS) like PsychoPy. The tutorials [fNIRS Stroop Experiment with PhysiolabXR and PsychoPy](https://physiolabxrdocs.readthedocs.io/en/latest/PsychoPy.html) and [PhysioLabXR + Unity P300 Speller Tutorial](https://physiolabxrdocs.readthedocs.io/en/latest/PhysioLabXRP300SpellerDemo.html) demonstrate how to use PsychoPy and Unity as SPS along with PhysioLabXR. Specifically, PhysioLabXR can make use of event markers with all the data streams for real-time data analysis and send analysis/inference results back to the SPS to give user feedback during the experiment. For example, in the latter example, the scripting interface sends the event markers indicating the decoded target character back to Unity, which in turn shows the participants what they spelled.

> >2- It seems that LSL has been used for data recording from any sensor. If so, the system performance depends on the network and the other software for reading data from a sensor. Could you please make it more clear or discuss it?
> Thanks for pointing this out! To clarify, PhysioLabXR is not currently designed as a stimulus presentation software (SPS) like PsychoPy. The tutorials [fNIRS Stroop Experiment with PhysiolabXR and PsychoPy](https://physiolabxrdocs.readthedocs.io/en/latest/PsychoPy.html) and [PhysioLabXR + Unity P300 Speller Tutorial](https://physiolabxrdocs.readthedocs.io/en/latest/PhysioLabXRP300SpellerDemo.html) demonstrate how to use PsychoPy and Unity as SPS along with PhysioLabXR. Specifically, PhysioLabXR can make use of event markers with all the data streams for real-time data analysis and send analysis/inference results back to the SPS to give user feedback during the experiment. For example, in the latter example, the scripting interface sends the event markers indicating the decoded target character back to Unity, which in turn shows the participants what they spelled.
>
We update (benchmark)[https://physiolabxrdocs.readthedocs.io/en/latest/Benchmarks.html] page to address the concern with network streaming speed. For devices that are not natively supported and have LSL or ZMQ support, users can opt to use these network streaming interfaces. Indeed, softwares such as Octopus-sensing, OpenViBE have data workers on separate processes and nevertheless data is shared between processes through LAN for visualization and other uses. PhysioLabXR handles this a little differently through multi-threading as noted in the “Software Design Principles” of the paper.

As for the performance concerns, the problem is two-folds:
1) pulling the data may slow other components of the software down: the data workers live on the main process, whereas other heavy-lifting processes such as replay and scripting are separate processes. Their running time should not be inference with each other.
2) delay on the network may skew the timestamp: external devices or software using LSL/ZMQ will generate timestamps when data is pushed. This ensures that the timestamp remains accurate even if the pulling of the data is delayed.

As an additional points, we added resource to show how to their customized data sources with [LSL](https://physiolabxrdocs.readthedocs.io/en/latest/DataStreamAPI/LSLStream.html) and  [ZMQ](https://physiolabxrdocs.readthedocs.io/en/latest/DataStreamAPI/ZMQStream.html#) in Python, Unity, or the scripting interface.
