### 2023-12-18 e7db22a
fix nothing received empty list bug
