### 2023-12-18 #5854 @mstimberg
@ApocalyVec , @HaowenWeiJohn Many thanks for your update and the detailed response.

:wave: @nastaran62 could you check whether the changes/replies address all your concerns and if possible provide more details with respect to your [installation issue](https://github.com/openjournals/joss-reviews/issues/5854#issuecomment-1858654957)
:wave: @lucask07 Given that you checked all the boxes of your checklist, can I assume that you are happy with the software/paper in its current state?
Thanks everyone :pray: !
