### 2023-12-19 b97f15f
changes adapt pyqtgraph
