+++ physiolabxr/examples/WriteYourOwnDataSourceExamples/LSLExampleOutlet.py
@@ -4,9 +4,10 @@ from random import random as rand
 
 # create a new stream info and outlet
 
-stream_name = 'python_lsl_my_stream_name'
+stream_name = 'stream'
+stream_type = 'LSL'
 
-n_channels = 8
+n_channels = 500
 # using the local_clock() to track elapsed time
 start_time = local_clock()
 # track how many samples we have sent
@@ -14,7 +15,7 @@ sent_samples = 0
 # set the sampling rate to 100 Hz
 nominal_sampling_rate = 100
 
-info = StreamInfo('stream_name', 'my_stream_type', n_channels, nominal_sampling_rate, 'float32',
+info = StreamInfo(stream_name, stream_type, n_channels, nominal_sampling_rate, 'float32',
                   'my_stream_id')
 outlet = StreamOutlet(info)
 
