+++ physiolabxr/ui/GroupPlotWidget.py
@@ -100,18 +100,28 @@ class GroupPlotWidget(QtWidgets.QWidget):
         distinct_colors = get_distinct_colors(len(channel_indices))
         self.legends = self.linechart_widget.addLegend()
         # self.linechart_widget.enableAutoRange(enable=False)
+        pens = []
+        names = []
         for channel_index_in_group, (channel_index, channel_name) in enumerate(
                 zip(channel_indices, self.channel_names)):
-            is_channel_shown = is_channels_shown[channel_index_in_group]
-            channel_plot_item = self.linechart_widget.plot([], [], pen=pg.mkPen(color=distinct_colors[channel_index_in_group]), name=channel_name)
-            self.channel_index_channel_dict[int(channel_index)] = channel_plot_item
-            if not is_channel_shown:
-                channel_plot_item.hide()  # TODO does disable do what it should do: uncheck from the plots
-            downsample_method = 'mean' if self.sampling_rate > AppConfigs().downsample_method_mean_sr_threshold else 'subsample'
-            channel_plot_item.setDownsampling(auto=True, method=downsample_method)
-            channel_plot_item.setClipToView(True)
-            channel_plot_item.setSkipFiniteCheck(True)
-            self.channel_plot_item_dict[channel_name] = channel_plot_item
+            # is_channel_shown = is_channels_shown[channel_index_in_group]
+            pens.append(pg.mkPen(color=distinct_colors[channel_index_in_group]))
+            names.append(channel_name)
+            # channel_plot_item = self.linechart_widget.plot([], [], pen=pg.mkPen(color=distinct_colors[channel_index_in_group]), name=channel_name)
+            # self.channel_index_channel_dict[int(channel_index)] = channel_plot_item
+            # if not is_channel_shown:
+            #     channel_plot_item.hide()  # TODO does disable do what it should do: uncheck from the plots
+            # downsample_method = 'mean' if self.sampling_rate > AppConfigs().downsample_method_mean_sr_threshold else 'subsample'
+            # channel_plot_item.setDownsampling(auto=True, method=downsample_method)
+            # channel_plot_item.setClipToView(True)
+            # channel_plot_item.setSkipFiniteCheck(True)
+            # self.channel_plot_item_dict[channel_name] = channel_plot_item
+        channel_plot_item = self.linechart_widget.plot([], [], pen=pens, name=names)
+        downsample_method = 'mean' if self.sampling_rate > AppConfigs().downsample_method_mean_sr_threshold else 'subsample'
+        channel_plot_item.setDownsampling(auto=True, method=downsample_method)
+        # channel_plot_item.setClipToView(True)
+        channel_plot_item.setSkipFiniteCheck(True)
+        # self.channel_plot_item_dict[channel_name] = channel_plot_item
 
     def init_image(self):
         self.plot_widget = pg.PlotWidget()
@@ -193,12 +203,17 @@ class GroupPlotWidget(QtWidgets.QWidget):
             # if line_chat_config.channels_constant_offset!=0:
             #     data = data +
 
-
             time_vector = np.linspace(0., duration, data.shape[1])
+            y_vals = np.zeros((len(channel_indices), time_vector.shape[0]))
             for index_in_group, channel_index in enumerate(channel_indices):
-                plot_data_item = self.linechart_widget.plotItem.curves[index_in_group]
-                if plot_data_item.isVisible():
-                    plot_data_item.setData(time_vector, data[channel_index, :]+linechart_config.channels_constant_offset*index_in_group)
+                y_vals[channel_index] = data[channel_index, :]+linechart_config.channels_constant_offset*index_in_group
+            plot_data_item = self.linechart_widget.plotItem.curves[0]
+            plot_data_item.setData(time_vector, y_vals)
[…]
