+++ physiolabxr/utils/video_capture_utils.py
@@ -38,8 +38,8 @@ def get_working_camera_ports():
             # SplashLoadingTextNotifier().set_loading_text("Video port %s is not working." %dev_port)
         else:
             is_reading, img = camera.read()
-            h, w, ch = img.shape
             if is_reading:
+                h, w, ch = img.shape
                 # SplashLoadingTextNotifier().set_loading_text("Video port %s is working and reads images (%s x %s)" %(dev_port,h,w))
                 working_cams.append({'stream_name': f"Camera {dev_port}", 'width': w, 'height': h, 'nchannels': ch, 'video_id': dev_port})
             else:
