### 2023-12-19 d3bdc5b
remove prent statement in click and dwell time
