### 2023-12-19 16dacfd
change channel number of LSLExampleOutlet.py
