### 2023-12-19 b15caa5
update requirements pyqtgraph to @wutwasthat's fork
