+++ requirements.txt
@@ -5,7 +5,7 @@ pyserial
 pylsl
 scikit-learn
 scipy~=1.11.0
-pyqtgraph
+git+https://github.com/wutwasthat/pyqtgraph.git@physio
 pyxdf
 pyscreeze
 opencv-python
