### 2023-12-20 98e4cdc
fix issue with creating/editing groups
