+++ physiolabxr/ui/GroupPlotWidget.py
@@ -102,8 +102,7 @@ class GroupPlotWidget(QtWidgets.QWidget):
         # self.linechart_widget.enableAutoRange(enable=False)
         pens = []
         names = []
-        for channel_index_in_group, (channel_index, channel_name) in enumerate(
-                zip(channel_indices, self.channel_names)):
+        for channel_index_in_group, (channel_index, channel_name) in enumerate(zip(channel_indices, self.channel_names)):
             # is_channel_shown = is_channels_shown[channel_index_in_group]
             pens.append(pg.mkPen(color=distinct_colors[channel_index_in_group]))
             names.append(channel_name)
@@ -204,11 +203,11 @@ class GroupPlotWidget(QtWidgets.QWidget):
             #     data = data +
 
             time_vector = np.linspace(0., duration, data.shape[1])
-            y_vals = np.zeros((len(channel_indices), time_vector.shape[0]))
-            for index_in_group, channel_index in enumerate(channel_indices):
-                y_vals[channel_index] = data[channel_index, :]+linechart_config.channels_constant_offset*index_in_group
-            plot_data_item = self.linechart_widget.plotItem.curves[0]
-            plot_data_item.setData(time_vector, y_vals)
+
+            y_vals = data[channel_indices]
+            channel_offsets = np.arange(y_vals.shape[0]) * linechart_config.channels_constant_offset
+            y_vals = y_vals + channel_offsets.reshape(-1, 1)
+            self.linechart_widget.plotItem.curves[0].setData(time_vector, y_vals)
                 # plot_data_item = self.linechart_widget.plotItem.curves[index_in_group]
                 # if plot_data_item.isVisible():
                 #     print('plotting channel', channel_index, 'in group', self.group_name)
