### 2023-12-20 19aeec6
fix issue with changing channel names
