+++ physiolabxr/scripting/physio/epochs.py
@@ -1,12 +1,15 @@
 import copy
 import os
+from collections import defaultdict
 from typing import Union, Dict, List
 
 import numpy as np
 import scipy
+from scipy.signal import spectrogram
 
 
-def get_event_locked_data(event_marker, data, events_of_interest, tmin, tmax, srate, return_last_event_time=False, verbose=None, event_channel=0):
+def get_event_locked_data(event_marker, data, events_of_interest, tmin, tmax, srate, return_last_event_time=False,
+                          verbose=None, event_channel=0, reject=None):
     """
     this function is used to get event locked data from a single modality or multiple modalities
 
@@ -16,6 +19,7 @@ def get_event_locked_data(event_marker, data, events_of_interest, tmin, tmax, sr
     @param tmin: time before event marker to include in the epoch
     @param tmax: time after event marker to include in the epoch
     @param srate: sampling rate of the datar
+    @param reject: int, reject the epoch with peak-to-peak amplitude greater than this value
     @param return_last_event_time: whether to return the time of the last found in the data
     @param verbose: whether to print out the number of events found for each event marker
 
@@ -32,7 +36,7 @@ def get_event_locked_data(event_marker, data, events_of_interest, tmin, tmax, sr
             _tmax = tmax[modality]
             _data = v
             _srate = srate[modality]
-            args = (event_marker, event_channel, _data, events_of_interest, _tmin, _tmax, _srate, True, verbose)
+            args = (event_marker, event_channel, _data, events_of_interest, _tmin, _tmax, _srate, True, verbose, reject)
             _locked_data, latest_event_start_time = _get_event_locked_data(*args)
             locked_data_modalities[modality] = _locked_data
 
@@ -46,7 +50,7 @@ def get_event_locked_data(event_marker, data, events_of_interest, tmin, tmax, sr
         return _get_event_locked_data(event_marker, event_channel, data, events_of_interest, tmin, tmax, srate, return_last_event_time, verbose)
 
 
-def _get_event_locked_data(event_marker, event_channel, data, events_of_interest, tmin, tmax, srate, return_last_event_time=False, verbose=None):
+def _get_event_locked_data(event_marker, event_channel, data, events_of_interest, tmin, tmax, srate, return_last_event_time=False, verbose=None, reject=None):
     """
     @param event_marker: tuple of event marker and its timestamps
     @param data: tuple of data and its timestamps
@@ -63,18 +67,23 @@ def _get_event_locked_data(event_marker, event_channel, data, events_of_interest
     rtn = {e: [] for e in events_of_interest}
     latest_event_start_time = -1
     epoch_length = int((tmax - tmin) * srate)
+    reject_count = defaultdict(int)
     for e in events_of_interest:
         this_event_marker_time = event_marker_time[event_marker == e]
         data_event_starts = [np.argmin(abs(data_time - (s+tmin))) for s in this_event_marker_time]
         data_event_ends = [epoch_length + s for s in data_event_starts]
         for i, j, e_time in zip(data_event_starts, data_event_ends, this_event_marker_time):
-            if j < len(data_time):
+            if j < len(data_time):  # if the epoch is not cut off by the end of the data
+                if reject is not None:
+                    if np.max(np.max(data[:, i:j], axis=0) - np.min(data[:, i:j], axis=0)) > reject:
+                        reject_count[e] += 1
+                        continue
[…]
