### 2023-12-20 a04feec
update generate event locked epochs function
