### 2023-12-20 ed5d625
update generate event locked epochs function
