### 2023-12-21 3de5e97
fix args for get_event_locked_data
