### 2023-12-22 f18d86a
user study test gaussian attention
