### 2023-12-22 #5854 @ApocalyVec
Hi @mstimberg,  @lucask07 and @nastaran62 ,

Just wanted to drop a quick note of heartfelt thanks to all the reviewers and the editor for your insightful and constructive feedback on our paper, "PhysioLabXR." Your thorough reviews and thoughtful suggestions have significantly contributed to enhancing both our manuscript, the software, and the docs in just a few month. We deeply appreciate the time and effort you've invested in helping us refine our work.

@mstimberg Please let us know if anything comes up in the submission process 👍

Happy holidays 🎇 and looking forward to crossing paths in future projects and research!

Best,
Leo & the PhysioLabXR Team 🚀
