### 2023-12-23 edef8a5
user study test gaussian attention
