### 2024-01-05 01714bd
Fix citation syntax issues
Avoids closing parentheses directly followed by opening parentheses
