+++ paper/paper.md
@@ -44,15 +44,15 @@ signal processing modules, support for recording and replaying experiments, and 
 
 
 
-*PhysioLabXR* is a Python-based open-source software platform for developing experiments for 
+*PhysioLabXR* is a Python-based open-source software platform for developing experiments for
 neuroscience and human–computer interaction (HCI)
 that involve real-time and multi-modal
-physiological data processing and interactive interfaces. *PhysioLabXR* provides native support for data sources such as electrophysiological sensors (e.g., EEG, EMG, and EOG), fNIRS, eye trackers, cameras, microphones, and screen capture, and implements the popular data transfer protocols Lab Streaming Layer (LSL) [@kothe2014labstreaminglayer] and ZeroMQ (ZMQ) [@zeromq]. It features multi-stream visualization methods, real-time digital signal 
-processing (DSP) modules, support for recording and replay experiments, and a Python-based scripting interface 
-for creating custom pipelines. 
+physiological data processing and interactive interfaces. *PhysioLabXR* provides native support for data sources such as electrophysiological sensors (e.g., EEG, EMG, and EOG), fNIRS, eye trackers, cameras, microphones, and screen capture, and implements the popular data transfer protocols Lab Streaming Layer [LSL\; @kothe2014labstreaminglayer] and ZeroMQ [ZMQ\; @zeromq]. It features multi-stream visualization methods, real-time digital signal
+processing (DSP) modules, support for recording and replay experiments, and a Python-based scripting interface
+for creating custom pipelines.
 
 *PhysioLabXR* has an architecture optimized through concurrency and parallelism to ensure efficient performance. We provide a set of detailed tutorials covering all features and example applications, such as a P300 speller with a Unity frontend [@Unity] and a mental arithmetic experiment interfacing with PsychoPy [@peirce2007psychopy]. An accompanying set of benchmarks demonstrates the ability of *PhysioLabXR* to handle high-throughput and
-multi-stream data reliably and efficiently. Published use cases show its versatility for VR and screen-based experiments [@lapborisuth2023pupil] [@koorathota2023multimodal] and sensor fusion studies  [@wei2022indexpen] [^1].
+multi-stream data reliably and efficiently. Published use cases show its versatility for VR and screen-based experiments [@lapborisuth2023pupil; @koorathota2023multimodal] and sensor fusion studies  [@wei2022indexpen] [^1].
 <!--
 The software reduces research cycle overhead and provides researchers with complete flexibility to build 
 customized systems. It represents an extensible framework to tackle the complexity of modern experiments at the intersection 
@@ -69,16 +69,16 @@ Recent years have seen a growing interest in multi-modal experiments, often invo
 Many emerging paradigms have found new roots in extended reality (XR) environments, including virtual reality (VR) and augmented reality (AR).
 Such experiments are increasingly fusing multiple modalities and combining different physiological measurements. For example, one sensor can generate events to extract
 meaningful data intervals from other sensors, such as fixation-related potential (FRP) studies in which EEG epochs are locked to visual fixations from eye trackers [@nikolaev2016combining]. Multiple physiological signals can also be combined to enhance their predictive
-power for use in applications ranging from emotion recognition [@koelstra2011deap] [@he2020advances] to movement actuation
+power for use in applications ranging from emotion recognition [@koelstra2011deap; @he2020advances] to movement actuation
 via sensorimotor rhythms [@sollfrank2016effect]. Further, multi-modal paradigms can facilitate the exploration of how different physiological systems interact; for example,
-pupil dilation can be used as a proxy for the locus coeruleus activity as measured via functional magnetic resonance imaging (fMRI) [@murphy2014pupil].
+pupil dilation can be used as a proxy for the locus coeruleus activity as measured via functional magnetic resonance imaging [fMRI\; @murphy2014pupil].
 
 <!--
[…]
