+++ physiolabxr/scripting/illumiRead/AOIAugmentationScript/AOIAugmentationScript.py
@@ -8,19 +8,19 @@ import pickle
 import sys
 import matplotlib.pyplot as plt
 import torch
-
+from eidl.utils.model_utils import get_subimage_model
 from pylsl import StreamInfo, StreamOutlet, cf_float32
 
-
-from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeData, GazeFilterFixationDetectionIVT, \
+from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationGazeUtils import GazeData, \
+    GazeFilterFixationDetectionIVT, \
     tobii_gaze_on_display_area_to_image_matrix_index, GazeType, gaze_point_on_image_valid
 from physiolabxr.scripting.illumiRead.AOIAugmentationScript.IntegrateAttention import integrate_attention
 from physiolabxr.scripting.RenaScript import RenaScript
 from physiolabxr.scripting.illumiRead.AOIAugmentationScript import AOIAugmentationConfig
 from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationUtils import *
-from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationConfig import EventMarkerLSLStreamInfo, GazeDataLSLStreamInfo
+from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationConfig import EventMarkerLSLStreamInfo, \
+    GazeDataLSLStreamInfo
 import torch
-
 
 
 class AOIAugmentationScript(RenaScript):
@@ -38,9 +38,7 @@ class AOIAugmentationScript(RenaScript):
 
         self.device = torch.device('cpu')
 
-
         self.current_image_name = None
-
 
         # self.vit_attention_matrix = ViTAttentionMatrix()
 
@@ -48,16 +46,24 @@ class AOIAugmentationScript(RenaScript):
 
         self.process_gaze_data_time_buffer = deque(maxlen=1000)
 
-        self.report_cleaned_image_info_dict = get_report_cleaned_image_info_dict(AOIAugmentationConfig.ReportCleanedImageInfoFilePath, merge_dict=True)
+        # self.report_cleaned_image_info_dict = get_report_cleaned_image_info_dict(AOIAugmentationConfig.ReportCleanedImageInfoFilePath, merge_dict=True)
+
+        if os.path.exists(AOIAugmentationConfig.SubImgaeHandlerFilePath):
+            with open(AOIAugmentationConfig.SubImgaeHandlerFilePath, 'rb') as f:
+                self.subimage_handler = pickle.load(f)
+        else:
+            self.subimage_handler = get_subimage_model()
+
+
+        self.current_image_info = ImageInfo()
 
         self.gaze_attention_matrix = GazeAttentionMatrix(device=self.device)
 
         # prevent null pointer in if statement been compiled before run
-        self.gaze_attention_matrix.set_image_shape(np.array([512,1024]))
-        self.gaze_attention_matrix.set_attention_patch_shape(np.array([16,32]))
-        test = self.gaze_attention_matrix.get_gaze_attention_grid_map(flatten=False)
+        # self.gaze_attention_matrix.set_image_shape(np.array([512, 1024]))
+        # self.gaze_attention_matrix.set_attention_patch_shape(np.array([16, 32]))
+        # test = self.gaze_attention_matrix.get_gaze_attention_grid_map(flatten=False)
 
-        self.current_image_info = ImageInfo()
         self.update_cue_now = False
 
         print("Experiment started")
@@ -65,7 +71,6 @@ class AOIAugmentationScript(RenaScript):
 
         # self.ivt_filter = GazeFilterFixationDetectionIVT(angular_speed_threshold_degree=100)
[…]
