+++ .github/workflows/unit-tests.yml
@@ -22,8 +22,6 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-      - uses: actions/checkout@v2
-      - uses: openrndr/setup-opengl@v1.1
       - name: Start Xvfb
         run: |
           sudo apt-get install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xinput0 libxcb-xfixes0
@@ -42,7 +40,8 @@ jobs:
           pip install pytest
           pip install pytest-qt
           sudo apt install gnome-screenshot
-
+      - uses: actions/checkout@v2
+      - uses: openrndr/setup-opengl@v1.1
       - name: Check job values
         run: |
            export QT_SELECT=qt6
