### 2024-01-08 50ba162
sudo apt-get install freeglut3 freeglut3-dev
