### 2024-01-08 3f13471
DEBUG MODE: add print statements
