### 2024-01-08 173a6a1
DEBUG MODE: print out startup_loadprest
