### 2024-01-08 085eff7
DEBUG MODE: pyqtgraph.setConfigOptions(useNumba=True, useOpenGL=False)
