+++ physiolabxr/startup/startup.py
@@ -62,7 +62,7 @@ def load_settings(revert_to_default=True, reload_presets=True, reload_configs=Tr
     # instantiate the GlabalSignals singleton object
     GlobalSignals()
     print("!!!GlobalSignals()")
-    pyqtgraph.setConfigOptions(useNumba=True, useOpenGL=True)
+    pyqtgraph.setConfigOptions(useNumba=True, useOpenGL=False)
     print("!!!pyqtgraph.setConfigOptions(useNumba=True, useOpenGL=True)")
 
 def load_ui_shared():
