### 2024-01-08 bddfd1c
DEBUG MODE: comment out pyqtgraph
