### 2024-01-08 b6bc052
DEBUG MODE: print out GroupEntry.py
