### 2024-01-08 3eb2321
DEBUG MODE: print out plotconfi
