### 2024-01-08 db5e833
DEBUG MODE: comment out cmap
