### 2024-01-08 79950f1
ready for pilot, add heatmap zmq and image rgba information
