+++ physiolabxr/examples/ZMQExampleOutlet.py
@@ -13,7 +13,7 @@ from pylsl import local_clock
 def main():
     topic = "CamCapture"
     srate = 120
-    port = "5557"
+    port = "5558"
 
     c_channels = 3
     width = 64
