+++ physiolabxr/scripting/illumiRead/AOIAugmentationScript/examples/read_image.py
@@ -1,0 +1,5 @@
+import cv2
+import numpy as np
+import matplotlib.pyplot as plt
+
+image = cv2.imread('bgra_image.png', cv2.IMREAD_UNCHANGED)
