+++ physiolabxr/scripting/illumiRead/AOIAugmentationScript/AOIAugmentationScript.py
@@ -1,12 +1,14 @@
 import time
 from collections import deque
 import cv2
+import numpy
 import numpy as np
 import time
 import os
 import pickle
 import sys
 import matplotlib.pyplot as plt
+import pylsl
 import torch
 from eidl.utils.model_utils import get_subimage_model
 from pylsl import StreamInfo, StreamOutlet, cf_float32
@@ -21,7 +23,8 @@ from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationUtils
 from physiolabxr.scripting.illumiRead.AOIAugmentationScript.AOIAugmentationConfig import EventMarkerLSLStreamInfo, \
     GazeDataLSLStreamInfo
 import torch
-
+import zmq
+from PIL import Image
 
 class AOIAugmentationScript(RenaScript):
     def __init__(self, *args, **kwargs):
@@ -60,7 +63,7 @@ class AOIAugmentationScript(RenaScript):
         self.gaze_attention_matrix = GazeAttentionMatrix(device=self.device)
 
         # prevent null pointer in if statement been compiled before run
-        # self.gaze_attention_matrix.set_image_shape(np.array([512, 1024]))
+        self.gaze_attention_matrix.set_maximum_image_shape(np.array([3000, 6000]))
         # self.gaze_attention_matrix.set_attention_patch_shape(np.array([16, 32]))
         # test = self.gaze_attention_matrix.get_gaze_attention_grid_map(flatten=False)
 
@@ -106,8 +109,14 @@ class AOIAugmentationScript(RenaScript):
             aoi_augmentation_attention_heatmap_lsl_outlet_info)
 
         # ################################################################################################################
+        # init heatmap zmq
+        self.aoi_augmentation_attention_heatmap_zmq_context = zmq.Context()
+        self.aoi_augmentation_attention_heatmap_zmq_socket = self.aoi_augmentation_attention_heatmap_zmq_context.socket(zmq.PUB)
+        self.aoi_augmentation_attention_heatmap_zmq_socket.bind("tcp://*:5557")
 
 
+
+        # ################################################################################################################
         # self.cur_attention_human = None
         # Start will be called once when the run button is hit.
 
@@ -144,7 +153,8 @@ class AOIAugmentationScript(RenaScript):
             block_marker = event_marker[AOIAugmentationConfig.EventMarkerLSLStreamInfo.BlockChannelIndex]
             state_marker = event_marker[AOIAugmentationConfig.EventMarkerLSLStreamInfo.ExperimentStateChannelIndex]
             image_index_marker = event_marker[AOIAugmentationConfig.EventMarkerLSLStreamInfo.ImageIndexChannelIndex]
-            user_inputs_marker = event_marker[AOIAugmentationConfig.EventMarkerLSLStreamInfo.UserInputsChannelIndex]
+            update_visual_cue_marker = event_marker[AOIAugmentationConfig.EventMarkerLSLStreamInfo.UpdateVisualCueMarker]
+            toggle_visual_cue_visibility_marker = event_marker[AOIAugmentationConfig.EventMarkerLSLStreamInfo.ToggleVisualCueVisibilityMarker]
 
             # ignore the block_marker <0 and state_marker <0 those means exit the current state
             if block_marker and block_marker > 0:  # evoke block change
@@ -188,6 +198,11 @@ class AOIAugmentationScript(RenaScript):
                         self.current_image_info.image_on_screen_shape = image_on_screen_shape
 
 
+                        # set image gaze attention matrix buffer
+                        self.gaze_attention_matrix.gaze_attention_pixel_map_buffer = torch.tensor(np.zeros(shape=self.current_image_info.original_image.shape[:2]),
[…]
