### 2024-01-08 1a1321f
Fix additional citation issues
