+++ paper/paper.bib
@@ -52,12 +52,13 @@
   note = {Accessed: Oct 28, 2023}
 }
 
-@misc{matlab,
-  title = {{MATLAB}},
-  note = {Version R2021b},
-  howpublished = {\url{https://www.mathworks.com/products/matlab.html}},
-  year = {2021},
-  publisher = {MathWorks},
+@software{matlab,
+year = {2021},
+author = {{MathWorks Inc.}},
+title = {MATLAB version: R2021b},
+publisher = {The MathWorks Inc.},
+address = {Natick, Massachusetts, United States},
+url = {https://www.mathworks.com}
 }
 
 @misc{pyqtgraph,
