### 2024-01-08 #5854 @mstimberg
@ApocalyVec @HaowenWeiJohn I made a PR with small formatting and citation fixes, please have a look. Could you also please take care of the points in the [checklist above](https://github.com/openjournals/joss-reviews/issues/5854#issuecomment-1876993775)? In particular, could you please make a release of the software and archive it (e.g. on [Zenodo](https://zenodo.org)), and then post the version number and the DOI here? Thanks!
