### 2024-01-08 #5854 @ApocalyVec
Hey @mstimberg,

Thanks for checking in. Here!

The newest release is v1.1.1 with changes from the review. [Link here](https://github.com/PhysioLabXR/PhysioLabXR/releases/tag/v1.1.0).

Release published on [Zenodo](https://zenodo.org/records/10471500). DOI: 10.5281/zenodo.10471500

We checked the changes you made to the paper. They all look good. Merged!

Thanks for your support throughout this review process. We appreciate the time and effort you and the reviewers have put into ensuring the quality and accuracy of our work! Please let me know if there's anything else needed from my side.

Cheers,
Leo & the PhysioLabXR Team
