### 2024-01-08 807c4be
update interactive mode TODO: add perceptual attention feedback
