### 2024-01-09 #5854 @mstimberg
@ApocalyVec The archive looks good to me in general, but there are a few minor issues that would be worth fixing (you can fix all this on Zenodo directly, without issuing a new version/DOI):
- For consistency with the paper, I'd suggest adding a DOI for co-author Steven Feiner
- The license is listed as "BSD 1-clause" on Zenodo, but it should be "BSD 3-clause"
- (really minor): the description is formatted as pre-formatted text, making it hard to read, in particular due to the very long lines
