### 2024-01-09 #5854 @mstimberg
:wave: @ApocalyVec @HaowenWeiJohn This is now *almost* ready to go. I've opened yet another [very small PR](https://github.com/PhysioLabXR/PhysioLabXR/pull/127) to add a DOI for the Octopus Sensing reference (thanks @editorialbot :bow:). Please fix/merge this and the minor issues in the Zenodo data [mentioned above](https://github.com/openjournals/joss-reviews/issues/5854#issuecomment-1882677844). **Note that it is not necessary to do a new release and/or Zenodo archive for these changes**.

Please give me a ping when you've handled these issues so that I can hand it over to the track editor for the final publication process.
