### 2024-01-09 #5854 @ApocalyVec
> @editorialbot set 10.5281/zenodo.10471500 as archive
>
> @ApocalyVec The archive looks good to me in general, but there are a few minor issues that would be worth fixing (you can fix all this on Zenodo directly, without issuing a new version/DOI):
>
> * For consistency with the paper, I'd suggest adding a DOI for co-author Steven Feiner
> * The license is listed as "BSD 1-clause" on Zenodo, but it should be "BSD 3-clause"
> * (really minor): the description is formatted as pre-formatted text, making it hard to read, in particular due to the very long lines

@mstimberg really appreciate the attention to details. All fixed on Zenodo:

- added ORCID for Steve Feiner
- updated the license to "BSD 3-Clause "New" or "Revised" License"
- reformatted description to plain text

One other change, I added ORCID for another author both in the paper and on Zenodo. Now everyone has ORCID attached.
