### 2024-01-11 #5854 @ApocalyVec
> @ApocalyVec As AEiC I will now help process this work for acceptance in JOSS. I have checked the paper, this review, your repository, and the archive link. Most seems in order. Below is one minor point that needs your attention.
>
> * [ ]  Please add the city, state to your affiliation, and please spell out USA as United States of America.

Updated! See [here](https://github.com/PhysioLabXR/PhysioLabXR/commit/6616b25cd05bb3eb0c2fbfe6303391e54818a6b6)

 * [x]  Please add the city, state to your affiliation, and please spell out USA as United States of America.
