+++ physiolabxr/scripting/generate_experiment_pipeline.py
@@ -1,0 +1,23 @@
+import numpy as np
+import random
+
+report_list = ['8898_OD_2021_widefield_report', '8902_OS_2021_widefield_report', '8914_OD_2021_widefield_report', '8918_OS_2021_widefield_report', '8919_OD_2021_widefield_report', '8924_OS_2021_widefield_report', '8935_OD_2021_widefield_report', '8939_OS_2021_widefield_report', '8974_OD_2021_widefield_report', '8981_OS_2021_widefield_report', '8984_OD_2021_widefield_report', '8988_OS_2021_widefield_report', '9000_OD_2021_widefield_report', '9002_OS_2021_widefield_report', '9025_OD_2021_widefield_report', '9071_OD_2021_widefield_report', '9075_OD_2021_widefield_report', '9078_OS_2021_widefield_report', '9108_OD_2021_widefield_report', '9111_OS_2021_widefield_report', '9140_OD_2021_widefield_report', '9172_OD_2021_widefield_report', '9175_OS_2021_widefield_report', '9183_OD_2021_widefield_report', '9184_OS_2021_widefield_report', '9238_OD_2021_widefield_report', 'RLS_001_OD_TC', 'RLS_001_OS_TC', 'RLS_004_OD_TC', 'RLS_004_OS_TC', 'RLS_005_OD_TC', 'RLS_005_OS_TC', 'RLS_006_OD_TC', 'RLS_008_OD_TC', 'RLS_008_OS_TC', 'RLS_011_OD_TC', 'RLS_011_OS_TC', 'RLS_012_OS_TC', 'RLS_016_OD_TC', 'RLS_018_OD_TC', 'RLS_018_OS_TC', 'RLS_023_OD_TC', 'RLS_023_OS_TC', 'RLS_024_OD_TC', 'RLS_024_OS_TC', 'RLS_025_OD_TC', 'RLS_028_OS_TC', 'RLS_031_OD_TC', 'RLS_031_OS_TC', 'RLS_033_OS_TC', 'RLS_036_OD_TC', 'RLS_036_OS_TC', 'RLS_039_OD_TC', 'RLS_039_OS_TC', 'RLS_040_OD_TC', 'RLS_041_OD_TC', 'RLS_042_OD_TC', 'RLS_043_OD_TC', 'RLS_043_OS_TC', 'RLS_044_OS_TC', 'RLS_045_OD_TC', 'RLS_045_OS_TC', 'RLS_046_OS_TC', 'RLS_048_OD_TC', 'RLS_048_OS_TC', 'RLS_050_OS_TC', 'RLS_054_OD_TC', 'RLS_054_OS_TC', 'RLS_058_OD_TC', 'RLS_059_OS_TC', 'RLS_060_OD_TC', 'RLS_060_OS_TC', 'RLS_062_OD_TC', 'RLS_062_OS_TC', 'RLS_063_OD_TC', 'RLS_067_OD_TC', 'RLS_068_OD_TC', 'RLS_068_OS_TC', 'RLS_073_OD_TC', 'RLS_073_OS_TC', 'RLS_077_OD_TC', 'RLS_077_OS_TC', 'RLS_079_OD_TC', 'RLS_079_OS_TC', 'RLS_082_OD_TC', 'RLS_082_OS_TC', 'RLS_083_OD_TC', 'RLS_083_OS_TC', 'RLS_084_OD_TC', 'RLS_084_OS_TC', 'RLS_085_OD_TC', 'RLS_086_OD_TC', 'RLS_086_OS_TC', 'RLS_087_OD_TC', 'RLS_087_OS_TC', 'RLS_089_OS_TC', 'RLS_091_OD_
[…]
