### 2024-01-12 #5854 @ApocalyVec
Hi [@mstimberg], [@lucask07], [@nastaran62], and the JOSS team,

We are thrilled to see "PhysioLabXR" published in JOSS! 🎉 We want to express our sincere gratitude to each of you for your invaluable contributions throughout the review process.

[@mstimberg] for your guidance as the editor in shaping the final version of our paper. To [@lucask07] and [@nastaran62], for your thorough and insightful reviews were incredibly helpful for both the paper and the software!

This collaborative effort has been a rewarding journey, and your support has been a significant part of its success. I'm looking forward to possibly crossing paths with y'all again in the future.

Again, thank you all once for your time, expertise, and constructive feedback!

Best regards,
[@ApocalyVec] & the PhysioLabXR Team
